==> part-1-1.qmd <==
# Population changes {#part-1-1}

This chapter covers population changes across the Partnership for Housing Affordability's main coverage area, including the City of Richmond and the counties of Chesterfield, Hanover, and Henrico.

```{r}
#| label: setup

library(tidycensus)
library(tidyverse)
library(janitor)
library(tigris)
library(sf)
library(sp)
library(scales)
library(ggiraph)
library(hdatools)
library(htmlwidgets)

rr <- c("51085", "51760", "51075", "51145", "51087", "51127", "51036", "51041")

pha <- c("51085", "51760", "51087", "51041")

rr_names <- c("Richmond city", "Hanover County", "Henrico County", "Chesterfield County",
              "Powhatan County", "Goochland County", "New Kent County", "Charles City County")

pha_names <- c("Richmond city", "Hanover County", "Henrico County", "Chesterfield County")

```

## Total population growth

The Richmond region has continued to grow between 2016 and 2020---adding a net of 41,457 residents across the four major localities. The most populous locality, Chesterfield County, experienced a near eight percent increase in its population during this timeframe.

```{r}
#| label: total-pop-data
#| eval: false

pep_total_raw <- get_estimates(
  geography = "county",
  state = "VA",
  variables = "POP",
  year = 2019,
  time_series = TRUE
)

census_raw <- get_decennial(
  geography = "county",
  state = "VA",
  year = 2020,
  sumfile = "pl",
  variables = "P1_001N"
)

projections <- read_csv("data/uva_proj.csv") |> 
  select(FIPS, `2030`, `2040`, `2050`) |> 
  pivot_longer(cols = 2:4,
               names_to = "year",
               values_to = "value") |> 
  mutate(counttype = case_when(
    year == "2030" ~ "Forecast",
    year == "2040" ~ "Forecast",
    year == "2050" ~ "Forecast"
  )) |> 
  rename(GEOID = FIPS) |> 
  mutate(GEOID = as.character(GEOID))

pep_total_clean <- pep_total_raw |>
  filter(!DATE %in% c(2, 3)) |> # Remove non-decennial 2010 counts
  mutate(year = # Translate date codes into years
    case_when(
      DATE == 1 ~ "2010",
      DATE == 4 ~ "2011",
      DATE == 5 ~ "2012",
      DATE == 6 ~ "2013",
      DATE == 7 ~ "2014",
      DATE == 8 ~ "2015",
      DATE == 9 ~ "2016",
      DATE == 10 ~ "2017",
      DATE == 11 ~ "2018",
      DATE == 12 ~ "2019")) |> 
  mutate(counttype = # Add descriptions to count types
      case_when(
        DATE == 1 ~ "Census population",
        TRUE ~ "Population estimate")) |> 
  select( # Simplify columns
    GEOID,
    year,
    counttype,
    value
  )

# Prep total population counts from 2020 Census summary file

census_clean <- census_raw |> 
  mutate(year = "2020", # Add year and count type columns
         counttype = "Census population") |>
  select( # Simplify columns
    GEOID,
    year,
    counttype,
    value
  )

# Append 2020 Census counts to PEP time series

population_data <- pep_total_clean |> 
  bind_rows(census_clean, projections)

rr_population_data <- population_data |> 
  subset(GEOID %in% rr)

rr_pop <- counties("VA", year = 2021) |>
  right_join(rr_population_data, by = "GEOID") |>
  select(NAMELSAD, year, counttype, value) |> 
  st_drop_geometry()

write_rds(rr_pop, "data/rr_pop.rds")

```

```{r}
#| label: fig-total-pop-plot
#| fig.cap: "Percent change in population by locality"

pha_pop <- read_rds("data/rr_pop.rds") |> 
  filter(NAMELSAD %in% pha_names)

pha_pop_compare <- pha_pop |> 
  filter(year == 2016 | year == 2020) |> 
  select(NAMELSAD, year, value) |> 
  pivot_wider(names_from = year,
              values_from = value) |> 
  mutate(difference = `2020` - `2016`,
         perchange = (difference/`2016`))

total_pop_plot <- ggplot(pha_pop_compare,
       aes(x = NAMELSAD,
           y = perchange,
           fill = NAMELSAD,
           label = percent(perchange),
           data_id = perchange,
           tooltip = percent(perchange))) +
  geom_col() +
  geom_col_interactive(size = 2) +
  scale_y_continuous(labels = percent_format()) +
  labs(title="Percent change in population by locality",
       subtitle="2016 to 2020",
       caption="**Source:** U.S. Census Bureau, Population Estimates Program.") +
  scale_fill_pha() +
  theme_pha() +
  add_zero_line("y")

total_pop_widget <- girafe(ggobj = total_pop_plot,
                           height_svg = 4)

if (knitr::is_html_output()) {
  
  total_pop_widget
  
} else {
  
  total_pop_plot +
    geom_text(vjust = 1.5,
              color ="white",
              size = 4,
              fontface ="bold")
  
  }
  
```

```{r}
#| label: widgets
#| eval: false

# Export widget via htmlwidgets

htmlwidgets::saveWidget(total_pop_widget, "widgets/htmlwidgets/total_pop.html", selfcontained = TRUE)

# Export widget via htmltools

htmltools::save_html(total_pop_widget, "widgets/htmltools/total_pop.html")

```


## Components of population change

In recent years, nearly two thirds of growth could be attributed to either domestic or international migration into the region. But between 2020 and 2021, that share increased to over three quarters---reducing the portion of the population growing due to natural increase to only 13 percent.

The region's growth continues to be driven primarily by new people coming from other parts of the state and nation (64 percent of growth between 2020 and 2021).

```{r}
#| label: components-data
#| eval: FALSE

# PEP does not have data available for 2020 and 2021. There is a gap between 2019 and 2021 because Census uses 2020 as a baseline. 

region <- c(51036, 51041, 51075, 51085, 51087, 51127, 51145, 51760)

pep_change_raw <- get_estimates(
  geography = "county",
  state = "VA",
  variables = c("NATURALINC", "DOMESTICMIG", "INTERNATIONALMIG"),
  time_series = TRUE
)

pep_change_clean <- pep_change_raw |> 
  mutate(GEOID = as.character(GEOID)) |> 
  filter(GEOID %in% rr) |> 
  mutate(year = # Translate date codes into years
    case_when(
      PERIOD == 1 ~ "2010",
      PERIOD == 2 ~ "2011",
      PERIOD == 3 ~ "2012",
      PERIOD == 4 ~ "2013",
      PERIOD == 5 ~ "2014",
      PERIOD == 6 ~ "2015",
      PERIOD == 7 ~ "2016",
      PERIOD == 8 ~ "2017",
      PERIOD == 9 ~ "2018",
      PERIOD == 10 ~ "2019")) |>
  mutate(component = # Rename components of change
    case_when(
      variable == "NATURALINC" ~ "Natural increase",
      variable == "DOMESTICMIG" ~ "Domestic migration",
      variable == "INTERNATIONALMIG" ~ "International migration")) |> 
  select( # Simplify columns
    GEOID,
    year,
    component,
    value
  )

# Manually add in 2021

pep_2021 <- read_csv("data/comp_chang_2021.csv") |> 
  pivot_longer(4:6,
               names_to = "component",
               values_to = "value") |> 
  mutate(year = 2021,
         component =
    case_when(
      component == "NATURALINC" ~ "Natural increase",
      component == "DOMESTICMIG" ~ "Domestic migration",
      component == "INTERNATIONALMIG" ~ "International migration")) |> 
  select(GEOID, year, component, value) |> 
  mutate(GEOID = as.character(GEOID))

rr_comp_change <- pep_change_clean |> 
  rbind(pep_2021) |>
  filter(GEOID %in% rr) |> 
  right_join(counties("VA", year = 2021), by = "GEOID") |>
  select(NAMELSAD, year, component, value)

write_rds(pep_comp_change, "data/rr_comp_change.rds")

```

```{r}
#| label: fig-components-plot
#| fig.cap: "Components of population change"

pha_comp_change <- read_rds("data/rr_comp_change.rds") |> 
  filter(NAMELSAD %in% pha_names)

comp_chg_summary <- pha_comp_change |> 
  group_by(year, component) |> 
  summarise(value = sum(value)) |> 
  mutate(year = as.numeric(year),
         percent = value/sum(value)) |> 
  filter(year >= 2016 & year != 2020) |> 
  mutate(year = as.character(year))

comp_chg_plot <- ggplot(comp_chg_summary,
       aes(x = year,
           y = percent,
           fill = component,
           data_id = percent,
           tooltip = percent(percent))) + 
  geom_col(position = "stack") +
  geom_col_interactive(position = "stack", size = 2) +
  scale_y_continuous(labels = percent_format()) +
  labs(title="Components of population change",
       subtitle = "Share of regional population growth by source",
       caption = "**Note:** Data not available for 2020.<br>**Source:** U.S. Census Bureau, Population Estimates Program.") +
  scale_fill_pha() +
  theme_pha() +
  theme(legend.position = "right") +
  add_zero_line("y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = comp_chg_plot,
         height_svg = 4) 
  
} else {
  
  comp_chg_plot
  
  }

```

## Population projections

Between 2020 and 2050, the region is expected to grow by nearly a third (29 percent)---reaching 1,338,306 residents.

```{r}
#| label: fig-forecasts-region-plot
#| fig.cap: "Population projections"

forecasts <- pha_pop |> 
  filter(year %in% c("2020", "2030", "2040", "2050")) |> 
  group_by(year, counttype) |> 
  summarise(value = sum(value))

forecasts_plot <- ggplot(forecasts,
       aes(x = year,
           y = value,
           fill = counttype,
           label = label_comma()(value),
           data_id = value,
           tooltip = label_comma()(value))) +
  geom_col() +
  geom_col_interactive() +
  scale_y_continuous(labels = label_comma(),
                     breaks = c(250000, 500000, 750000, 1000000, 1250000)) +
  labs(title = "Population projections",
       subtitle = "Regional population in 2020 and forecast to 2050",
       fill = "Estimate type",
       caption = "**Source:** University of Virginia Weldon Cooper Center for Public Service.") +
  scale_fill_pha() +
  theme_pha() +
  theme(
    legend.position = "right",
    legend.title = element_text(hjust = 0)
  ) +
  add_zero_line("y")

# Difference between 2050 and 2020 is 302,780. 29 percent increase.

if (knitr::is_html_output()) {
  
  girafe(ggobj = forecasts_plot,
         height_svg = 4) 
  
} else {
  
  forecasts_plot +
    geom_text(vjust = 1.5,
              color ="white",
              size = 3,
              fontface ="bold")
  
  }

```

Over the next 30 years, Chesterfield County will continue to lead growth across the region. By 2050, Chesterfield is expected to surpass half a million residents, growing by 38 percent from the 2020 Census estimates.

Population growth trends will largely continue as they have with Hanover County experiencing the second greatest growth from their 2020 estimates (27 percent increase). Henrico County follows with a 26 percent increase (+88,565), while the City of Richmond will only increase by about a fifth (20 percent) over 30 years.

```{r}
#| label: fig-forecasts-locality-plot
#| fig.cap: "Population projections by locality"

forecasts_locality <- pha_pop |> 
  filter(year %in% c("2020", "2030", "2040", "2050")) |> 
  group_by(NAMELSAD, year, counttype) |> 
  summarise(value = sum(value))

forecasts_locality_plot <- ggplot(forecasts_locality,
       aes(x = year,
           y = value,
           fill = NAMELSAD,
           data_id = value,
           tooltip = label_comma()(value))) +
  geom_col() +
  geom_col_interactive(size = 2) +
  facet_wrap(~NAMELSAD) +
  scale_y_continuous(labels = label_comma()) +
  labs(title = "Population projections by locality",
       subtitle = "Local population in 2020 and forecast to 2050",
       caption = "**Source:** University of Virginia Weldon Cooper Center for Public Service.") +
  scale_fill_pha() +
  theme_pha() +
  add_zero_line("y")

# Chesterfield - 140,266 difference / 38 percent increase
# Hanover - 30,134 difference / 27 percent increase
# Henrico - 88,565 difference / 26 percent increase
# Richmond - 43,815 difference / 19 percent increase

if (knitr::is_html_output()) {
  
  girafe(ggobj = forecasts_locality_plot,
         height_svg = 4) 
  
} else {
  
  forecasts_locality_plot
  
  }

```
