==> part-1-3.qmd <==
# Incomes and wages {#part-1-3}

```{r}
#| label: setup

library(tidyverse)
library(tidycensus)
library(tigris)
library(lubridate)
library(readxl)
library(janitor)
library(leaflet)
library(scales)
library(ggiraph)
library(ggtext)
library(hdatools)

# rr includes Charles City, Goochland, New Kent, and Powhatan.

rr <- c("51085", "51760", "51075", "51145", "51087", "51127", "51036", "51041")

pha <- c("51085", "51760", "51087", "51041")

# Set palette colors; may need to find a palette that better complements the PHA branding.

rr_names <- c("Richmond city", "Hanover County", "Henrico County", "Chesterfield County",
              "Powhatan County", "Goochland County", "New Kent County", "Charles City County")

pha_names <- c("Richmond city", "Hanover County", "Henrico County", "Chesterfield County")

tenure <- c("Homeowner", "Renter")

years <- 2016:2020

```

This chapter covers the changes in incomes and wages among households and occupations across the region. 

## Household incomes

### Incomes by tenure

From 2016 to 2020, the region saw large increases in the number of six-figure income households, particularly among homeowners (well over 25,000), but also renters (almost 6,500). This growth can likely be attributed to both new high-income residents from outside the region, as well as income growth among households already in the region.

There was also a minor increase in the number of middle-income renters earning between \$50,000 and \$100,000, reflecting continued demand for new market-rate apartments---along with affordable starter homes.

```{r}
#| label: hh-inc-data
#| eval: FALSE

b25118_vars <- load_variables(2020, "acs5") |> 
  filter(str_sub(name, end = 6) %in% "B25118")

b25118_raw <- map_dfr(years, function(yr){
  b25118_pull <- get_acs(
    geography = "county",
    state = "VA",
    table = "B25118",
    year = yr,
    survey = "acs5",
    cache_table = TRUE
  ) |> 
    mutate(year = yr)
  })

b25118_raw <- b25118_raw |> 
  subset(GEOID %in% rr)

b25118_vars_cleaned <- b25118_vars |> 
  separate(label, into = c("est", "total", "tenure", "income"), sep = "!!") |>  
  select(variable = name, tenure, income) |> 
  mutate(tenure = case_when(
    tenure == "Owner occupied:" ~ "Homeowner",
    tenure == "Renter occupied:" ~ "Renter"
  )) |> 
  drop_na()

b25118_data <- b25118_raw |> 
  right_join(b25118_vars_cleaned, by = "variable") |> 
  select(NAME, year, tenure, income, estimate) |> 
  mutate(NAME = str_remove_all(NAME, ", Virginia")) |>
  mutate(income = case_when(
    income == "Less than $5,000" ~ "Less than $15,000",
    income == "$5,000 to $9,999" ~ "Less than $15,000",
    income == "$10,000 to $14,999" ~ "Less than $15,000",
    income == "$15,000 to $19,999" ~ "$15,000 to $24,999",
    income == "$20,000 to $24,999" ~ "$15,000 to $24,999",
    income == "$25,000 to $34,999" ~ "$25,000 to $49,999",
    income == "$35,000 to $49,999" ~ "$25,000 to $49,999",
    income == "$50,000 to $74,999" ~ "$50,000 to $74,999",
    TRUE ~ income
  )) |> 
  group_by(NAME, year, tenure, income) |> 
  summarise(estimate = sum(estimate)) |> 
  filter(year %in% c(2016, 2020)) |> 
  pivot_wider(names_from = year,
              values_from = estimate) |>
  mutate(change = `2020` - `2016`,
         pct_change = change/`2016`)
  
write_rds(b25118_data, "data/b25118_data.rds")

```

```{r}
#| label: fig-hh-inc-plot
#| fig.cap: "Change households by tenure and income level"

b25118_data <- read_rds("data/b25118_data.rds") |> 
  filter(NAME %in% pha_names) |>
  group_by(tenure, income) |> 
  summarise(`2016` = sum(`2016`),
            `2020` = sum(`2020`)) |> 
  mutate(change = `2020`-`2016`,
         pct_change = change/`2016`)

hh_inc_plot <- ggplot(b25118_data, 
       aes(y = income,
           x = change,
           fill = tenure,
           data_id = change,
           tooltip = label_comma()(change))) +
  geom_col(position = "dodge") +
  geom_col_interactive(position = "dodge", size = 2) +
  #facet_grid(~tenure) +
  scale_x_continuous(labels = label_comma()) +
  scale_y_discrete(limits = c("Less than $15,000","$15,000 to $24,999",
                              "$25,000 to $49,999", "$50,000 to $74,999",
                              "$75,000 to $99,999", "$100,000 to $149,999",
                              "$150,000 or more")) +
                   #,labels = function(x) str_wrap(x, width = 12)
  labs(title = "Change households by tenure and income level",
       subtitle="2016 to 2020",
       fill = element_blank(),
       caption = "**Source:** U.S. Census Bureau, American Community Survey, Table B25118.") +
  scale_fill_pha() +
  theme_pha() +
  theme(legend.position = "right") +
  flip_gridlines() +
  add_zero_line("y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = hh_inc_plot,
         height_svg = 4)
  
} else {
  
  hh_inc_plot
  
  }

```

:::{.callout-important}

## Why use medians

Using [medians](https://www.investopedia.com/terms/m/median.asp) instead of averages is a standard data analysis practice because it accounts for outliers. An average household income or home sales price would be influenced adversely by one or a few data points at the high end --- causing data skewing. 

:::

Typical homeowner incomes continue to be well above average renter incomes across the region. When adjusted for inflation, incomes across tenures for each locality show very minor to modest growth. Incomes in the city---for both homeowners and renters---remain significantly below those in the surrounding counties. The median household income for homeowners in the counties is around three times that of renters in Richmond.

```{r}
#| label: med-income-data
#| eval: FALSE

b25119_vars <- load_variables(2020, "acs5") |> 
  filter(str_sub(name, end = 6) %in% "B25119")

b25119_raw <- map_dfr(years, function(yr){
  b25119_pull <- get_acs(
    geography = "county",
    state = "VA",
    table = "B25119",
    year = yr,
    survey = "acs5",
    cache_table = TRUE
  ) %>%
    mutate(year = yr)
  })

b25119_raw <- b25119_raw |> 
  subset(GEOID %in% rr)

b25119_vars_cleaned <- b25119_vars |> 
  separate(label, into = c("est", "income", "total", "tenure"), sep = "!!") |>  
  select(variable = name, tenure) |> 
  mutate(tenure = case_when(
    tenure == "Owner occupied (dollars)" ~ "Homeowner",
    tenure == "Renter occupied (dollars)" ~ "Renter"
  ))

b25119_vars_cleaned$tenure <- b25119_vars_cleaned$tenure |>  replace_na('All households')

b25119_data <- b25119_raw |> 
  right_join(b25119_vars_cleaned, by = "variable") |> 
  select(NAME, GEOID, year, tenure, estimate, moe)

cpi <- read_excel("data/CPI_U_RS.xlsx")
cpi <- cpi |> 
  rename(year = Year,
         priceindex = Index) |> 
  transform(year = as.numeric(year))

b25119_cpi <- b25119_data |> 
  left_join(cpi, by = 'year') |> 
  transform(dollars20 = ((381.2/priceindex)*estimate)) |> 
  select(NAME, GEOID, year, tenure, dollars20, cdollars = estimate) |> 
  mutate(NAME = str_remove_all(NAME, ", Virginia")) |>
  filter(tenure != "All households") |> 
  select(year, NAME, tenure, dollars20, cdollars)

write_rds(b25119_cpi, "data/b25119_cpi.rds")

```

```{r}
#| label: fig-med-income-plot
#| fig.cap: "Median houshold income by tenure and locality"

b25119_cpi <- read_rds("data/b25119_cpi.rds") |> 
  filter(NAME %in% pha_names)

med_inc_plot <- ggplot(b25119_cpi, 
       aes(x = year,
           y = dollars20,
           color = NAME)) +
  geom_line(stat = "identity", size = 1) +
  geom_point_interactive(
    aes(data_id = dollars20,
        tooltip = label_dollar()(dollars20),
        group = 1),
    size = 0.5) +
  facet_wrap(~tenure) +
  scale_color_pha() +
  scale_y_continuous(labels = dollar_format(), limits = c(0, NA)) +
  labs(title = "Median houshold income by tenure and locality",
       subtitle = "2016 to 2020 | 2020 inflation-adjusted dollars",
       color = "Locality",
       caption = "**Source:** U.S. Census Bureau, American Community Survey, Table B25119.") +
  theme_pha() +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = med_inc_plot,
         height_svg = 4)
  
} else {
  
  med_inc_plot
  
  }

```

:::{.callout-important}

## Adjusting for inflation

When comparing dollar figures at different points in time, it is important to account for the effects of [inflation](https://www.investopedia.com/ask/answers/042415/what-impact-does-inflation-have-time-value-money.asp) on the dollar value. The value of a dollar in 2020 is not the same as in 2010 because inflation increases 

To accurately compare data like median household incomes across time, we must convert all dollar values to the same timeframe. The [Consumer Price Index (CPI)](https://www.investopedia.com/terms/c/consumerpriceindex.asp) is used to adjust previous dollar figures to the most recent data point. 

:::
### Incomes by race and ethnicity

Typical incomes in the region remain unequal by race and ethnicity. Households with the highest incomes include Asian and white, non-Hispanic residents in the counties---earning well above \$75,000. Black and Hispanic households consistently have the lowest median incomes, along with multiracial households in Henrico and Richmond.

```{r}
#| label: inc-race-data
#| eval: FALSE

b19013 <- paste0("B19013", LETTERS[2:9])

b19013_defns <- load_variables(2020, "acs5") %>%
  filter(str_sub(name, end = 7) %in% b19013) %>%
  filter(str_detect(name, "PR") == FALSE)

concept_to_race <- function(x) {
  out <- x %>%
    str_remove_all("MEDIAN HOUSEHOLD INCOME IN THE PAST 12 MONTHS \\(\\IN 2020 INFLATION-ADJUSTED DOLLARS\\)\\ \\(|\\)") %>%
    str_to_title()
}

b19013_cleaned <- b19013_defns %>%
  mutate(race = concept_to_race(concept)) %>%
  separate(label, c("estimate", "medhhincome"), sep = "!!") %>%
  select(variable = name, medhhincome, race) %>%
  mutate(across(.fns = ~replace_na(.x, "All")),
          across(.fns = ~str_remove_all(.x, ":")))

b19013_raw <- map_dfr(b19013, function(tb) {
  yearly_data <- map_dfr(years, function(yr) {

    acs_pull <- get_acs(
      geography = "county",
      table = tb,
      year = yr,
      state = "VA"
    ) %>%
    left_join(b19013_cleaned, by = "variable")

    acs_rearranged <- acs_pull %>%
      mutate(year = yr) %>%
      select(variable, year, NAME, GEOID, race, medhhincome,
             estimate, moe)

    acs_rearranged
  })
  yearly_data
})

cpi <- read_excel("data/CPI_U_RS.xlsx") |> 
  rename(year = Year,
         priceindex = Index) |> 
  transform(year = as.numeric(year))

b19013_data <- b19013_raw |> 
  subset(GEOID %in% rr) |> 
  mutate(across(.fns = ~str_remove_all(.x, "Alone Householder")),
         across(.fns = ~str_remove_all(.x, ", Virginia")),
         across(.fns = ~trimws(.x))) |> 
  select(year, NAME, GEOID, race, estimate, moe) |> 
  transform(year = as.numeric(year),
            estimate = as.numeric(estimate))|> 
  left_join(cpi, by = 'year')|> 
  select(year, NAME, GEOID, race, estimate, moe, priceindex) |> 
  transform(dollars20 = ((381.2/priceindex)*estimate)) |> 
  mutate(race = case_when(
    race == "Black Or African American" ~ "Black",
    race == "Two Or More Races Householder" ~ "Multiracial",
    race == "White Alone, Not Hispanic Or Latino Householder" ~ "White, non-Hispanic",
    race == "Hispanic Or Latino Householder" ~ "Hispanic or Latino",
    TRUE ~ race
  )) |> 
  filter(race %in% c("White, non-Hispanic", "Black", "Asian", "Multiracial", "Hispanic or Latino"))

write_rds(b19013_data, "data/b19013_data.rds")

```

```{r}
#| label: fig-inc-race-plot
#| fig.cap: "Median household income by race and ethnicity"

b19013_data <- read_rds("data/b19013_data.rds") |> 
  filter(NAME %in% pha_names) |> 
  mutate(cv = ((as.numeric(moe)/1.645)/estimate)) |> 
  filter(cv < .15)

inc_race_plot <- ggplot(b19013_data,
       aes(x = year,
           y = dollars20/1000,
           color = race)) +
  geom_line(stat = "identity", size = 1) +
  geom_point_interactive(
    aes(data_id = dollars20,
        tooltip = label_dollar()(dollars20),
        group = 1),
    size = 0.5) +
  facet_wrap(~NAME, nrow = 1) +
  scale_color_pha() +
  scale_y_continuous(labels = label_dollar(suffix = "k"),
                     limits = c(0, NA),
                     breaks = c(0, 25, 50, 75, 100, 125)) +
  labs(title = "Median household income by race and ethnicity",
       subtitle = "2016 to 2020 | 2020 inflation-adjusted dollars",
       caption = "**Note:** Estimates with low reliability are omitted.<br>**Source:** U.S. Census Bureau, American Community Survey, Table B19013.") +
  theme_pha() +
  theme(legend.position = "bottom",
        legend.key = element_blank(),
        axis.text.x = element_text(angle = 90, vjust = 0.5)) +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = inc_race_plot,
         height_svg = 4)
  
} else {
  
  inc_race_plot
  
  }

```

### Incomes by family type

Household incomes also vary by the presence of children or other related individuals. Throughout the region, non-family households (i.e., persons living alone or with unrelated persons) consistently have typical incomes below \$50,000. In Henrico and Chesterfield counties, families living with and without children under 18 have very similar income levels. This trend is different in Hanover, where families with children have much higher incomes, as well as Richmond, where they have much lower incomes.

```{r}
#| label: inc-child-data
#| eval: FALSE

# Table B19125: Median Family Income by Presence of Own Children

b19125_vars <- load_variables(2020, "acs5") |> 
  filter(str_sub(name, end = 6) %in% "B19125")

b19125_raw <- map_dfr(years, function(yr){
  b25119_pull <- get_acs(
    geography = "county",
    state = "VA",
    table = "B19125",
    year = yr,
    survey = "acs5",
    cache_table = TRUE
  ) %>%
    mutate(year = yr)
  })

b19125_raw <- b19125_raw |> 
  filter(GEOID %in% rr)

b19125_vars_cleaned <- b19125_vars |> 
  separate(label, into = c("est", "total", "type"), sep = "!!") |>  
  select(variable = name, type) |> 
  mutate(type = case_when(
    type == "With own children of the householder under 18 years" ~ "Family, with children",
    type == "No own children of the householder under 18 years" ~ "Family, no children"
  )) |> 
  drop_na()

b19125_data <- b19125_raw |> 
  right_join(b19125_vars_cleaned, by = "variable") |> 
  select(NAME, GEOID, year, type, estimate)

# Table B19202: Median Nonfamily Household Income

b19202_raw <- map_dfr(years, function(yr){
  b19202_pull <- get_acs(
    geography = "county",
    state = "VA",
    variable = "B19202_001",
    year = yr,
    survey = "acs5",
    cache_table = TRUE
  ) %>%
    mutate(year = yr)
  })

b19202_data <- b19202_raw |> 
  filter(GEOID %in% rr) |> 
  mutate(type = "Non-family") |> 
  select(NAME, GEOID, year, type, estimate)

cpi <- read_excel("data/CPI_U_RS.xlsx") |> 
  rename(year = Year,
         priceindex = Index) |> 
  transform(year = as.numeric(year))

b19125_b19202_data <- b19125_data |> 
  bind_rows(b19202_data) |> 
  mutate(NAME = str_remove_all(NAME, ", Virginia"))

b19125_b19202_cpi <- b19125_b19202_data |> 
  left_join(cpi, by = 'year') |> 
  transform(dollars20 = ((381.2/priceindex)*estimate)) |> 
  select(year, NAME, type, dollars20)

write_rds(b19125_b19202_cpi, "data/b19125_b19202_cpi.rds")

```

```{r}
#| label: fig-inc-child-plot
#| fig.cap: "Median household income by family type"

b19125_b19202_cpi <-  read_rds("data/b19125_b19202_cpi.rds") |> 
  filter(NAME %in% pha_names)

inc_child_plot <- ggplot(b19125_b19202_cpi,
       aes(x = year,
           y = dollars20/1000,
           color = type)) +
  facet_wrap(~NAME, nrow = 1) +
  geom_line(stat = "identity", size = 1) +
  geom_point_interactive(
    aes(data_id = dollars20,
        tooltip = label_dollar()(dollars20),
        group = 1),
    size = 0.5) +
  scale_color_pha() +
  scale_y_continuous(labels = label_dollar(suffix = "k"),
                     limits = c(0, NA),
                     breaks = c(0, 25, 50, 75, 100, 125)) +
  labs(title = "Median household income by family type",
       subtitle = "2016 to 2020 | 2020 inflation-adjusted dollars",
       caption = "**Source:** U.S. Census Bureau, American Community Survey, Tables B19125 and B19202.") +
  theme_pha() +
  theme(legend.position = "bottom",
        legend.key = element_blank(),
        axis.text.x = element_text(angle = 90, vjust = 0.5)) +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = inc_child_plot,
         height_svg = 4)
  
} else {
  
  inc_child_plot
  
  }

```

## Wages

::: callout-note
Wage data in this section is sourced from the Occupational Employment and Wage Statistics (OEWS) program of the Bureau of Labor Statistics. OEWS is updated annually, most recently for 2021 data. This dataset provides a rich look into wage distribution by industry and occupation. 

However, OEWS is only available at the national, state, and metro levels. Therefore, the data below covers the full Richmond, Virginia Metropolitan Statistical Area (MSA) rather than the (smaller) PHA region.
:::

### Wage change by percentile

While regional wages increased across the board from May 2019 to May 2021, the largest percent increases in average wages were among jobs that paid at and below the median wage. In fact, the largest growth occurred in the lowest 10th percentile of wages, due in large part to state lawmakers adopting incremental increases to Virginia's minimum wage in 2020. The first increase from \$7.25 to \$9.50 per hour took effect in 2021.

::: callout-note
Today, state minimum wage is \$11.00 per hour. Under [current law](https://lis.virginia.gov/cgi-bin/legp604.exe?201+sum+SB7), it will increase again to \$12.00 in 2023. Lawmakers must reenact the measure by July 2024 to initiative further increases to \$15.00 per hour by 2026.
:::

Another factor in this low-end wage growth is likely the [increased pay](https://www.bls.gov/opub/ted/2022/24-percent-of-establishments-increased-pay-or-paid-bonuses-because-of-covid-19-pandemic.htm) offered by many businesses, especially in the food, retail, and accommodation sectors, to encourage workers to return during the COVID-19 recovery.

```{r}
#| label: wage-pct-data
#| eval: FALSE

# Load OEWS files already filtered to Richmond, VA MSA

oews_2019 <- read_xlsx("data/oews_rva_2019.xlsx",
                       sheet = "m2019",
                       na = c("*", "**", "#"))

oews_2021 <- read_xlsx("data/oews_rva_2021.xlsx",
                       sheet = "m2021",
                       na = c("*", "**", "#")) |> 
  janitor::clean_names(case = "snake")

# Pull out annual wage percentiles for each year

percentiles_2019 <- oews_2019 |> 
  filter(occ_code == "00-0000") |> 
  select(9, 24:28) |> 
  mutate(year = "y2019", .before = 1)

percentiles_2021 <- oews_2021 |> 
  filter(occ_code == "00-0000") |> 
  select(10, 26:30) |> 
  mutate(year = "y2021", .before = 1)

# Combine years, rearrange, and calculate percent change over time

percentiles <- bind_rows(percentiles_2019, percentiles_2021) |> 
  pivot_longer(3:7) |> 
  pivot_wider(names_from = year,
              values_from = value) |> 
  mutate(pct_change = (y2021 - y2019)/y2019,
         wage = case_when(
           name == "a_pct10" ~ "10th percentile",
           name == "a_pct25" ~ "25th percentile",
           name == "a_median" ~ "Median",
           name == "a_pct75" ~ "75th percentile",
           name == "a_pct90" ~ "90th percentile"
         )) |> 
  mutate(wage = fct_relevel(wage,
                            "10th percentile",
                            "25th percentile",
                            "Median",
                            "75th percentile",
                            "90th percentile"))

write_rds(percentiles, "data/oews_pct.rds")

```

```{r}
#| label: fig-wage-pct-plot
#| fig.cap: "Percent change in annual wage in Richmond, VA MSA"

oews_pct <- read_rds("data/oews_pct.rds")

wage_pct_plot <- ggplot(oews_pct,
       aes(x = pct_change,
           y = wage,
           fill = wage,
           data_id = pct_change,
           tooltip = label_percent()(pct_change))) +
  geom_col() +
  geom_col_interactive(size = 2) +
  scale_x_continuous(labels = label_percent()) +
  scale_fill_pha() +
  labs(title = "Percent change in annual wage by wage percentile",
       subtitle = "Richmond, VA MSA | May 2019 to May 2021",
       caption = "**Source:** U.S. Bureau of Labor Statistics, Occupational Employment and Wage Statistics.") +
  theme_pha() +
  add_zero_line(axis = "x") +
  flip_gridlines()


if (knitr::is_html_output()) {
  
  girafe(ggobj = wage_pct_plot,
         height_svg = 4)
  
} else {
  
  wage_pct_plot +
    geom_text(aes(label = label_percent(accuracy = 0.1)(pct_change),
                  color = wage),
              nudge_x = 0.01) +
    scale_color_pha()
  
  }

```

### Wage change by occupation

Over this same period, wages in the region grew for four of the five most common occupation categories by total employment numbers. Workers in the Transportation and Material Moving sector saw the largest increases---from an average annual salary of \$30,250 to \$36,370 (over 20 percent).

Jobs in Food Preparation and Serving, Sales, and Business and Financial Operations sectors---totaling more than 162,000 workers in the region as of May 2021---also saw wage growth, but less than the 13.3 percent average increase. Meanwhile, wages among Office and Administrative Support positions remained nearly the same (-0.2 percent) from 2019 to 2021.

```{r}
#| label: wage-occ-data
#| eval: FALSE

# Find 5 most common major occupation categories from 2021

oews_occ <- oews_2021 |> 
  filter(o_group == "major") |> 
  slice_max(tot_emp, n = 5) |> 
  select(9, 10, tot_emp_2021 = 12, a_median_2021 = 28)

# Add 2019 wage data for those occupations

oews_occ <- oews_occ |> 
  left_join(oews_2019, by = "occ_code") |> 
  select(1, occ_title = 2, 3, 4, tot_emp_2019 = 14, a_median_2019 = 29)

# Calculate wage percent change and add median for all occupations

oews_occ <- oews_occ |> 
  mutate(pct_change = (a_median_2021 - a_median_2019) / a_median_2019) |> 
  add_row(occ_title = "All Occupations", pct_change = as.numeric(oews_pct[oews_pct$wage == "Median",5]))

write_rds(oews_occ, "data/oews_occ.rds")

```

```{r}
#| label: fig-wage-occ-plot
#| fig.cap: "Percent change in annual wage for top 5 most common occupation categories"

oews_occ <- read_rds("data/oews_occ.rds")

wage_occ_plot <- ggplot(oews_occ,
       aes(x = pct_change,
           y = reorder(occ_title, pct_change),
           fill = occ_title,
           data_id = pct_change,
           tooltip = label_percent()(pct_change))) +
  geom_col() +
  geom_col_interactive(size = 2) +
  scale_x_continuous(labels = label_percent(), limits = c(-0.005, 0.23)) +
  scale_y_discrete(labels = function(x) str_wrap(x, width = 30)) +
  scale_fill_pha() +
  labs(title = "Percent change in annual wage for<br>top 5 most common occupation categories",
       subtitle = "Richmond, VA MSA | May 2019 to May 2021",
       caption = "**Source:** U.S. Bureau of Labor Statistics, Occupational Employment and Wage Statistics.") +
  theme_pha() +
  add_zero_line(axis = "x") +
  flip_gridlines()

if (knitr::is_html_output()) {
  
  girafe(ggobj = wage_occ_plot,
         height_svg = 4)
  
} else {
  
  wage_occ_plot +
    geom_text(aes(label = label_percent(accuracy = 0.1)(pct_change),
                  color = occ_title),
              nudge_x = 0.015) +
    scale_color_pha()
  
  }

```
