==> part-1-4.qmd <==
# Special populations {#part-1-4}

```{r}
#| label: setup

library(tidyverse)
library(tidycensus)
library(sf)
library(leaflet)
library(scales)
library(ggtext)
library(ggiraph)
library(hdatools)

# rr includes Charles City, Goochland, New Kent, and Powhatan.

rr <- c("51085", "51760", "51075", "51145", "51087", "51127", "51036", "51041")

pha <- c("51085", "51760", "51087", "51041")

rr_names <- c("Richmond city", "Hanover County", "Henrico County", "Chesterfield County",
              "Powhatan County", "Goochland County", "New Kent County", "Charles City County")

pha_names <- c("Richmond city", "Hanover County", "Henrico County", "Chesterfield County")

years <- 2016:2020

```

This chapter covers trends among populations that experience a disability that may impact their ability to access and maintain housing.

## Independent living difficulty

In the American Community Survey (ACS), the Census Bureau collects a range of characteristics to capture the range of different disability types found in the population. One important disability type available in ACS data is *independent living difficulty*, which includes persons who:

> *Because of a physical, mental, or emotional problem, [have] difficulty doing errands alone such as visiting a doctor’s office or shopping.*

As a result, persons with these difficulties often face significant housing challenges as well.

### By age

From 2016 to 2020, the region added almost 2,600 more persons with independent living difficulties. The largest increases occurred among young adults under 35, as well as "young" seniors between 65 and 74. The latter group will see their needs increase acutely in the next decade as they continue to age and potentially become more dependent on others.

```{r}
#| label: ind-liv-age-data
#| eval: FALSE

# Table B18107 Sex by age by independent living difficulty

# "Because of a physical, mental, or emotional problem, having difficulty doing errands alone such as visiting a doctor’s office or shopping (DOUT)."

b18107_vars <- load_variables(2020, "acs5") |> 
  filter(str_sub(name, end = 6) %in% "B18107")

b18107_raw <- map_dfr(years, function(yr){
  b10107_pull <- get_acs(
    geography = "county",
    state = "VA",
    table = "B18107",
    year = yr,
    survey = "acs5",
    cache_table = TRUE
  ) |> 
    mutate(year = yr)
  })

b18107_raw <- b18107_raw |> 
  subset(GEOID %in% rr)

b18107_vars_cleaned <- b18107_vars |> 
  separate(label, into = c("est", "total", "sex", "age", "indliv"), sep = "!!") |> 
  select(variable = name, sex , age, indliv) |> 
  drop_na() |> 
  mutate(across(.fns = ~str_remove_all(.x, ":")))

b18107_data <- b18107_raw |> 
  right_join(b18107_vars_cleaned, by = "variable") |> 
  select(NAME, GEOID, year, sex, age, indliv, estimate) |> 
  mutate(NAME = str_remove_all(NAME, ", Virginia")) |>
  group_by(year, NAME, age, indliv) |> 
  summarise(estimate = sum(estimate)) |> 
  filter(indliv == "With an independent living difficulty") |> 
  pivot_wider(names_from = year,
              values_from =  estimate) |> 
  transform(change = `2020` - `2016`)

write_rds(b18107_data, "data/b18107_data.rds")

```

```{r}
#| label: fig-ind-liv-age-plot
#| fig.cap: "Net change in individuals with independent living difficulties by age"

b18107_data <- read_rds("data/b18107_data.rds") |> 
  filter(NAME %in% pha_names) |> 
  group_by(age) |> 
  summarise(change = sum(change))

ind_liv_age_plot <- ggplot(b18107_data,
       aes(x = age,
           y = change,
           fill = age,
           data_id = change,
           tooltip = label_comma()(change))) +
  geom_col_interactive(size = 2) +
  scale_y_continuous(labels = label_comma()) +
  scale_fill_pha() +
  labs(title = "Net change in individuals with<br>independent living difficulties by age",
       subtitle = "2016 to 2020",
       caption = "**Source:** U.S. Census Bureau, American Community Survey, Table B18107.") +
  theme_pha() +
  add_zero_line("y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = ind_liv_age_plot,
         height_svg = 4)
  
} else {
  
  ind_liv_age_plot
  
  }

```

### By tenure

::: callout-note
The detailed estimates for persons with independent living difficulties in this and the next section are not available from the standard ACS tables published by the Census Bureau. The data in these sections come from the Public Use Microdata Sample (PUMS), which are available only by special Public Use Microdata Areas (PUMAs) which contain at least 100,000 people.

While PUMA boundaries align with Chesterfield County, Henrico County, and Richmond city, the PUMA containing Hanover County also includes Powhatan, Goochland, New Kent, King William, Charles City counties.
:::

Nearly all persons with independent living difficulties throughout the region live in regular homes, and not assisted living facilities or other group quarters. Most are in homes that they own, or in homes owned by another occupant, such as a spouse. This is not the case in Richmond, however, where about half live in rented homes.

```{r}
#| label: ind-liv-tenure-data
#| eval: FALSE

# Load PUMA geographies for Virginia

pumas <- pumas(state = "VA")

# Filter PUMAs to PHA region

pumas_pha <- pumas |>
  filter(str_detect(NAMELSAD10, "Chesterfield|Hanover|Henrico|Richmond"))

indliv_pums <- get_pums(
  variables = c("TEN", "NP", "DOUT"),
  state = "VA",
  year = 2020,
  puma = pumas_pha$PUMACE10)

indliv_raw <- indliv_pums |> 
  select(SERIALNO, PUMA, SPORDER, WGTP, PWGTP, TEN, NP, DOUT) |> 
  filter(DOUT == "2") |> 
  mutate(geography = case_when(
    PUMA %in% c("04101", "04102", "04103") ~ "Chesterfield County",
    PUMA %in% c("51224", "51225") ~ "Henrico County",
    PUMA == "51235" ~ "Richmond city",
    PUMA == "51215" ~ "Remainder of PDC")) |> 
  mutate(TEN = case_when(
    TEN == "b" ~ "Group quarters",
    TEN %in% c("1", "2") ~ "Owner-occupied",
    TEN == "3" ~ "Renter-occupied",
    TRUE ~ "No payment provided")) |> 
  mutate(NP = case_when(
    NP == 1 ~ "Living alone",
    NP == 2 ~ "2 persons",
    NP == 3 ~ "3 persons",
    NP == 4 ~ "4 persons",
    TRUE ~ "5 or more persons"))

indliv_data <- indliv_raw |>
  select(geography, TEN, NP, WGTP, PWGTP)

write_rds(indliv_data, "data/indliv_data.rds")

```

```{r}
#| label: fig-ind-liv-tenure-plot
#| fig.cap: "Percent of persons with independent living difficulties by household tenure"

indliv_tenure <- read_rds("data/indliv_data.rds") |> 
  group_by(geography, TEN) |> 
  summarise(est = sum(PWGTP)) |> 
  mutate(pct = est/sum(est)) |> 
  mutate(TEN = fct_relevel(TEN, "Owner-occupied", "Renter-occupied", "Group quarters", "No payment provided"))

ind_liv_tenure_plot <- ggplot(indliv_tenure,
       aes(x = TEN,
           y = pct,
           fill = TEN,
           data_id = pct,
           tooltip = label_percent(accuracy = 0.1)(pct))) +
  facet_wrap(~geography, nrow = 1) +
  geom_col_interactive(size = 2) +
  scale_x_discrete() +
  scale_y_continuous(labels = label_percent()) +
  scale_fill_pha() +
  labs(title = "Percent of persons with independent living difficulties<br>by household tenure",
       subtitle = "Includes persons in group quarters",
       caption = "**Source:** U.S. Census Bureau, American Community Survey,<br>Public Use Microdata Sample, 2016-2020 5-year estimates.") +
  theme_pha() +
  add_zero_line("y") +
  theme(legend.position = "bottom",
        axis.text.x = element_blank())

if (knitr::is_html_output()) {
  
  girafe(ggobj = ind_liv_tenure_plot,
         height_svg = 4)
  
} else {
  
  ind_liv_tenure_plot
  
  }

```

### By household size

Persons with independent living difficulties are most likely to live with one other person in their home. Slightly larger households (3 to 4 persons total) are also common. Still, more than 15 percent live alone---including nearly one-in-four in Richmond. However, based on ACS data collection methods, "living alone" also includes persons residing in group quarters.

```{r}
#| label: fig-ind-liv-size-plot
#| fig.cap: "Percent of persons with independent living difficulties by household size"

indliv_size <- read_rds("data/indliv_data.rds") |> 
  group_by(geography, NP) |> 
  summarise(est = sum(PWGTP)) |> 
  mutate(pct = est/sum(est)) |> 
  mutate(NP = fct_relevel(NP, "Living alone"))

ind_liv_size_plot <- ggplot(indliv_size,
       aes(x = NP,
           y = pct,
           fill = NP,
           data_id = pct,
           tooltip = label_percent(accuracy = 0.1)(pct))) +
  facet_wrap(~geography, nrow = 1) +
  geom_col_interactive(size = 2) +
  scale_x_discrete() +
  scale_y_continuous(labels = label_percent()) +
  scale_fill_pha() +
  labs(title = "Percent of persons with independent living difficulties<br>by household size",
       subtitle = "Persons in group quarters are counted as living alone",
       caption = "**Source:** U.S. Census Bureau, American Community Survey,<br>Public Use Microdata Sample, 2016-2020 5-year estimates.") +
  theme_pha() +
  add_zero_line("y") +
  theme(legend.position = "bottom",
        axis.text.x = element_blank())

if (knitr::is_html_output()) {
  
  girafe(ggobj = ind_liv_size_plot,
         height_svg = 4)
  
} else {
  
  ind_liv_size_plot
  
  }

```

## Veterans with disabilities

Veterans of military service have access to a range of Department of Veterans Administration (VA) benefits, including VA home loans. These benefits also include disability payments for veterans with service-connected disabilities.

To award disability benefits, the VA assigns each disabled veteran a [rating](https://www.va.gov/disability/about-disability-ratings/) from zero to 100 percent based on the severity of their disability or disabilities. A higher rating reflects more significant impairments, and accordingly, additional paid benefits to cover lost wages and extra healthcare services.

From 2016 to 2020, the number of veterans in the region with a service-connected disability increased by more than 2,800. A significant majority of this growth occurred among veterans with disability rating of 70 percent or higher, or those with the most severe physical and/or mental health challenges.

Despite the increased benefits level associated with the higher rating, these disabled veterans may be challenged to find accessible and affordable housing options without additional support.

```{r}
#| label: vets-data
#| eval: FALSE

# Table B21100: Service-connected Disability-rating Status and Ratings for Civilian Veterans 18 Years and Over

b21100_vars <- load_variables(2020, "acs5") |> 
  filter(str_sub(name, end = 6) %in% "B21100")

b21100_raw <- map_dfr(years, function(yr){
  b21100_pull <- get_acs(
    geography = "county",
    state = "VA",
    table = "B21100",
    year = yr,
    survey = "acs5",
    cache_table = TRUE
  ) |> 
    mutate(year = yr)
  })

b21100_raw <- b21100_raw |> 
  subset(GEOID %in% rr)

b21100_vars_cleaned <- b21100_vars |> 
  separate(label, into = c("est", "total", "disability", "rating"), sep = "!!") |> 
  select(variable = name, rating) |> 
  drop_na()

b21100_data <- b21100_raw |> 
  right_join(b21100_vars_cleaned, by = "variable") |> 
  select(NAME, GEOID, year, rating, estimate) |> 
  mutate(NAME = str_remove_all(NAME, ", Virginia")) |>
  pivot_wider(names_from = year,
              values_from =  estimate) |> 
  transform(change = `2020` - `2016`)

write_rds(b21100_data, "data/b21100_data.rds")

```

```{r}
#| label: fig-vets-plot
#| fig.cap: "Net change in veterans with service-connected disability by disability rating"

b21100_data <- read_rds("data/b21100_data.rds") |> 
  filter(NAME %in% pha_names) |> 
  group_by(rating) |> 
  summarise(change = sum(change))
 
vets_plot <- ggplot(b21100_data,
       aes(x = rating,
           y = change,
           fill = rating,
           data_id = change,
           tooltip = label_comma()(change))) +
  geom_col_interactive(size = 2) +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 15)) +
  scale_y_continuous(labels = label_comma()) +
  scale_fill_pha() +
  labs(title = "Net change in veterans with<br>service-connected disability by disability rating",
       subtitle = "2016 to 2020",
       caption = "**Source:** U.S. Census Bureau, American Community Survey, Table B21100.") +
  theme_pha() +
  add_zero_line("y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = vets_plot,
         height_svg = 4)
  
} else {
  
  vets_plot
  
  }


```
