==> part-2-2.qmd <==
# Rental homes {#part-2-2}

```{r}
#| label: setup

library(tidyverse)
library(tidycensus)
library(janitor)
library(glue)
library(stringi)
library(zoo)
library(lubridate)
library(tigris)
library(ggthemes)
library(sf)
library(sp)
library(scales)
library(viridis)
library(ggiraph)
library(ggtext)
library(hdatools)

# rr includes Charles City, Goochland, New Kent, and Powhatan.

rr <- c("51085", "51760", "51075", "51145", "51087", "51127", "51036", "51041")
pha <- c("51085", "51760", "51087", "51041")

pha_names_costar <- c("Richmond City", "Henrico County", "Hanover County", "Chesterfield County")

submarkets <- c("Midlothian" = "#2ca25f",
                "Chesterfield County" = "#99d8c9",
                "Hanover County" = "#8856a7",
                "Western Henrico County" = "#f03b20",
                "Eastern Henrico County" = "#feb24c",
                "West End" = "#253494",
                "Downtown Richmond" = "#2c7fb8",
                "East End" = "#41b6c4",
                "South Richmond" = "#7fcdbb",
                "Northside" = "#c7e9b4")

```
This chapter covers trends in the rental housing market across the four main Partnership for Housing Affordability localities, including the City of Richmond and counties of Chesterfield, Hanover, and Henrico.

## Supply

### Change in stock

While many renters across the region do live in multifamily buildings (with 5 or more units), the second largest share of rental housing is single-family housing (either attached or detached). In 2020, over a third (37 percent) of rental housing in the region consisted of single-family housing, while 49 percent was located in buildings with 5 or more units. There has been little change in these percentages since 2016. 

Changes in the shares of rental housing have been small --- but those changes have been among rental housing with 20 or more units (17 percent in 2016 to 19 percent in 2020) and 2 to 4 unit buildings (14 percent in 2016 down to 13 percent in 2020). 

```{r}
#| label: fig-ro-structure-percent
#| fig.cap: "Change in share of renter-occupied housing units by structure type"

b25127_data <- read_rds("data/b25127_data.rds") |> 
  filter(GEOID %in% pha,
         tenure == "Renter") |> 
  ungroup()

ro_structure_percent <- b25127_data |> 
  filter(structure != "Other") |>
  select(5, 6, 10) |> 
  pivot_longer(
    cols = 2:3,
    names_to = "year",
    values_to = "estimate") |> 
  group_by(structure, year) |> 
  summarise(estimate = sum(estimate)) |> 
  group_by(year) |> 
  mutate(pct = estimate/sum(estimate))

ro_structure_plot <- ggplot(ro_structure_percent,
       aes(x = structure,
           y = pct,
           fill = year,
           label = label_percent(accuracy = 0.1)(pct),
           data_id = pct,
           tooltip = label_percent(accuracy = 0.1)(pct))) +
  geom_col(position = "dodge") +
  geom_col_interactive(position = "dodge", size = 2) +
  scale_y_continuous(labels = label_percent()) +
  scale_fill_pha() +
  labs(title = "Change in share of renter-occupied housing units by structure type",
       subtitle = "2016 to 2020",
       caption = "**Source:** U.S. Census Bureau, American Community Survey, Table B25127.") +
  theme_pha() +
  add_zero_line("y") +
  theme(legend.position = "right")

if (knitr::is_html_output()) {
  
  girafe(ggobj = ro_structure_plot,
         height_svg = 4) 
  
} else {
  
  ro_structure_plot +
    geom_text(vjust = 1.5,
              color = "white",
              position = position_dodge(width = 0.9))
  
  }

```

The raw changes in rental housing were most felt in Henrico County and the City of Richmond. In Henrico, there was a 1,930 increase in single-family rental housing and a 1,357 decrease in 2 to 4 unit rental housing (i.e. duplexes, triplexes, and quads).

The City of Richmond saw a contrasting decrease in single-family rentals (-1,921), while also experiencing a 2,134 increase in rental housing located in buildings with 20 or more units. Chesterfield County has seen slight increases in multifamily housing of all types, while Hanover County has not seen much change at all.

```{r}
#| label: fig-ro-structure
#| fig.cap: "Change in renter-occupied housing units by structure type"

ro_structure <- b25127_data |> 
  filter(structure != "Other") |> 
  select(NAME, structure, change) |> 
  group_by(NAME, structure) |> 
  summarise(change = sum(change))

ro_str_local_plot <- ggplot(ro_structure,
       aes(x = structure,
           y = change,
           fill = NAME,
           data_id = change,
           tooltip = label_comma()(change))) +
  facet_wrap(~NAME) +
  geom_col_interactive(position = "dodge", size = 2) +
  scale_y_continuous(labels = label_comma()) +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 10)) +
  scale_fill_pha() +
  labs(title = "Change in renter-occupied housing units by structure type",
       subtitle = "2016 to 2020",
      caption = "**Source:** U.S. Census Bureau, American Community Survey, Table B25127.") +
  theme_pha() +
  add_zero_line("y")
  
if (knitr::is_html_output()) {
  
  girafe(ggobj = ro_str_local_plot,
         height_svg = 4) 
  
} else {
  
  ro_str_local_plot
  
  }

```

### Age of stock

Since 2016, the region has seen major changes in the age of its rental stock as existing homes transition from being owned to leased out, or vice-versa. Of note, every locality except for Hanover saw significant increases in the number of renter-occupied homes built between 1980 and 1999.

These homes---now over 20 years old---are likely becoming the target of investors purchasing from homeowners, making certain improvements, and renting them out. In Henrico County, this trend was even more prevalent among homes built between 1960 and 1979.

:::{.callout-note}

## Market Value Analysis (MVA)

In 2021, Richmond Memorial Health Foundation (RMHF) and PlanRVA commissioned a second Market Value Analysis (MVA) of the Richmond region. The MVA is a "is a data-based, field-validated analysis and mapping of a community’s housing market." Richmond's MVA provides fine-grained data analysis of neighborhood changes and trends, including residential vacancy and investor sales.

Learn more about the MVA [here](https://rmhfoundation.org/resource/richmond-memorial-health-foundation-co-sponsored-analysis-of-the-richmond-areas-housing-market-reveals-important-trends-for-residents-policymakers-and-investors/).

:::

Conversely, Chesterfield and Henrico each had over 1,000 homes built between 2000 and 2009 change from renter- to owner-occupied. The largest losses in rental stock, however, occurred in Richmond among homes built prior to 1980. Several factors could explain this decline:

* Actual demolition of very old, low-quality homes,
* Duplexes and triplexes converted into single-family homes, and
* Single-family rentals purchased by buyers who now live in the home.

```{r}
#| label: fig-ro-year-built
#| fig.cap: "Change in renter-occupied housing units by year built"

ro_yrblt <- b25127_data |> 
  group_by(NAME, yrbuilt) |> 
  summarise(change = sum(change))
  
ro_age_plot <- ggplot(ro_yrblt,
       aes(x = yrbuilt,
           y = change,
           fill = NAME,
           data_id = change,
           tooltip = label_comma()(change))) +
  facet_wrap(~NAME) +
  geom_col_interactive(position = "dodge", size = 2) +
  scale_y_continuous(labels = label_comma()) +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 10)) +
  scale_fill_pha() +
  labs(title = "Change in renter-occupied housing units by year built",
       subtitle = "2016 to 2020",
       caption = "**Source:** U.S. Census Bureau, American Community Survey, Table B25127.") +
  theme_pha() +
  add_zero_line("y")
  
if (knitr::is_html_output()) {
  
  girafe(ggobj = ro_age_plot,
         height_svg = 4) 
  
} else {
  
  ro_age_plot
  
  }

```

### Bedrooms

Rental homes in the Richmond region are most likely to have one or two bedrooms. While the number of one-bedroom apartments has continued to increase (+1,617) from 2016, the number of two-bedroom units has decreased by 2,500. 

The increasing supply of one-bedroom apartments coincides with a similar increase in studio apartments---these unit sizes reflect new apartments, largely in Richmond, marketed for college students and other young adults.

The dwindling number of two-bedroom rental homes may reflect small single-family rentals in older neighborhoods transitioning to owner-occupancy, as there is a similar (but much less significant) decline in three-bedroom units.

```{r}
#| label: fig-ro-bedrooms
#| fig-cap: "Change in renter-occupied housing units by number of bedrooms"

b25042_data <- read_rds("data/b25042_data.rds") |> 
  filter(GEOID %in% pha,
         tenure == "Renter") |> 
  group_by(br) |> 
  summarise(change = sum(change))

ro_bed_plot <- ggplot(b25042_data,
       aes(x = br,
           y = change,
           fill = br,
           data_id = change,
           tooltip = label_comma()(change))) +
  geom_col_interactive(size = 2) + 
  scale_y_continuous(labels = label_comma()) +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 16)) +
  scale_fill_pha() +
  labs(title = "Change in renter-occupied housing units<br>by number of bedrooms",
       subtitle = "2016 to 2020",
       caption = "**Source:** U.S. Census Bureau, American Community Survey, Table B25042.") +
  theme_pha() +
  add_zero_line("y")
  

if (knitr::is_html_output()) {
  
  girafe(ggobj = ro_bed_plot,
         height_svg = 4) 
  
} else {
  
  ro_bed_plot
  
  }

```

### Production

Construction of multifamily properties (with 5 units or more) has been sporadic since the end of the Great Recession. In all localities aside from Hanover County, there have been waves and dips in the multifamily building construction. Hanover has seen little to no activity throughout the last two decades, while Chesterfield County and Richmond have seen the bulk of activity.

During the latter half of the last decade, Chesterfield County had a boom in multifamily construction --- nearing 1,500 units in 2019. Meanwhile, Richmond's multifamily construction saw dips following the Great Recession and again in 2018, but has largely been up in the last couple years of the 2010s. Although Henrico County had dips in 2016 and 2018, multifamily construction has more often than not been above the 700 unit mark.

```{r}
#| label: fig-mf-permits
#| fig.cap: "Multifamily building permits"

pha_cbps <- read_rds("data/rr_cbps.rds") |> 
  filter(GEOID %in% pha,
         type != "1-unit") |> 
  mutate(type = case_when(
    type != "5+ units" ~ "2-4 units",
    TRUE ~ type
  )) |> 
  group_by(NAME, year, type) |> 
  summarise(units = sum(units))

bps_plot <- ggplot(pha_cbps,
       aes(x = year, y = units, fill = type,
           data_id = units, tooltip = label_comma()(units))) +
  geom_col_interactive(size = 2) +
  facet_wrap(~NAME) +
  scale_y_continuous(labels = label_comma()) +
  scale_fill_pha() +
  labs(title = "Multifamily building permits",
       subtitle = "2000 to 2020",
       caption = "**Source:** U.S. Census Bureau, Building Permits Survey.") +
  theme_pha() +
  add_zero_line("y") +
  theme(legend.position = "right")

```

## Rental market

### Average market asking rent

Rental demand reached [a fever pitch amid the ongoing COVID-19 pandemic](https://richmond.com/news/local/rent-in-richmond-region-surged-during-the-pandemic-two-bedroom-apartments-average-1-340-a/article_751d018b-b5b4-5592-9582-9e825bda674b.html). With eviction moratoriums and a flow of rental assistance, low supply gave way to historic rent increases. The average market asking rent in the region reached a two-decade high of \$1,395 in the first quarter of 2022.

```{r}
#| label: fig-avg-rent
#| fig-cap: "Average asking rent"

# Aggregated average market asking rent per unit for Chesterfield, Hanover, Henrico, and Richmond City.

# CoStar Filters: Construction Status - Existing

# As of mid July 

pha_rent <- read_csv("data/pha_avg_rent.csv") |> 
  clean_names() |> 
  mutate(period = as.Date(as.yearqtr(period, format = "%Y Q%q"), frac = 1)) |> 
  pivot_longer(cols = 2:5,
               names_to = "type",
               values_to = "rent") |> 
  mutate(type = case_when(
    type == "rent_infl" ~ "Inflation-adjusted",
    type == "rent_cur" ~ "Nominal",
    type == "rent_infl_diff" ~ "Change from previous quarter (inflation-adjusted)",
    type == "rent_cur_diff" ~ "Change from previous quarter (nominal)"
  )) |> 
  mutate(year = year(period))
  
pha_rent_avg <- pha_rent |> 
  filter(type == "Inflation-adjusted" | type == "Nominal") |> 
  filter(between(period, as.Date("2019-03-31"), as.Date("2022-06-30"))) |> 
  mutate(period = as.yearqtr(period))

rent_avg_plot <- ggplot(pha_rent_avg,
       aes(x = period, y = rent, color = type)) +
  geom_line(size = 1) +
  geom_point_interactive(aes(data_id = rent, tooltip = label_dollar()(rent)),
                         size = 0.5) +
  scale_y_continuous(labels = label_dollar(),
                     n.breaks = 9) +
  scale_x_yearqtr(format = "%Y Q%q") +
  scale_color_pha() +
  labs(title = "Average asking rent",
       subtitle = "2019 Q1 to 2022 Q2",
       color = "Rent value",
       caption = "**Source:** CoStar Group, Inc.") +
  theme_pha() +
  theme(legend.position = "right")

if (knitr::is_html_output()) {
  
  girafe(ggobj = rent_avg_plot,
         height_svg = 4) 
  
} else {
  
  rent_avg_plot
  
  }

```

Large quarterly increases in average rents began in early 2021 and have continued to the present. From the first to second quarters of this year, rents increased by \$31. However, this relative growth was very near the change in inflation over that same period. 

```{r}
#| label: fig-avg-rent-change
#| fig-cap: "Quarterly change in average asking rent"

pha_rent_chg <- pha_rent |> 
  filter(str_detect(type, "Change")) |>
  filter(between(period, as.Date("2019-03-31"), as.Date("2022-06-30"))) |> 
  mutate(period = as.yearqtr(period))

rent_chg_plot <- ggplot(pha_rent_chg,
       aes(x = period,
           y = rent,
           fill = type,
           data_id = rent,
           tooltip = label_dollar()(rent))) +
  geom_col_interactive(size = 2) +
  facet_wrap(~type) +
  scale_y_continuous(labels = label_dollar(),
                     n.breaks = 9) +
  scale_x_yearqtr(format = "%Y Q%q") +
  scale_fill_pha() +
  labs(title = "Quarterly change in average asking rent",
       subtitle = "2019 Q1 to 2022 Q2",
       caption = "**Source:** CoStar Group, Inc.") +
  theme_pha() +
  add_zero_line("y")
  
if (knitr::is_html_output()) {
  
  girafe(ggobj = rent_chg_plot,
         height_svg = 4) 
  
} else {
  
  rent_chg_plot
  
  }

```

### Rents by submarket

Although not adjusted for inflation, rents by submarket show that there are distinct average rents across the region. Since 2010, the steepest increases have occurred in the counties. Northside Richmond remains the least expensive submarket with an average rent of \$1,037 in the second quarter of this year, while Midlothian is the most expensive at \$1,655.

```{r}
#| label: fig-avg-rent-submarket
#| fig.cap: "Average asking rent by submarket"

# Disaggregate by Richmond CoStar Submarket

# CoStar Filters: Construction Status - Existing

# Added locality field manually

rr_sub_rent <- read_csv("data/rr_submarket_data.csv") |> 
  clean_names() |> 
  separate(geography, into = c("region", "state", "submarket"), sep = " - ") |> 
  select(period, locality, submarket, "rent" = market_asking_rent_unit, vacancy_rate) |> 
  mutate(rent = parse_number(rent)) |> 
  mutate(period = as.Date(as.yearqtr(period, format = "%Y Q%q"), frac = 1)) |> 
  subset(locality %in% pha_names_costar) |> 
  group_by(locality) |> 
  mutate(submarket = fct_reorder(submarket, rent, .desc = TRUE)) |> 
  filter(between(period, as.Date("2010-03-31"), as.Date("2022-06-30"))) |> 
  mutate(period = as.yearqtr(period))

# Issue with above is that rents are not inflation-adjusted. May need to revisit. I've requested from CoStar to allow for an inflation-adjusted average rent per unit pull, BUT who knows if they can add it. Might consider just having a table with most recent rents.

sub_rent_plot <- ggplot(rr_sub_rent,
       aes(x = period,
           y = rent, 
           color = submarket)) +
  geom_line(size = 1) + 
  geom_point_interactive(aes(data_id = rent, tooltip = label_dollar()(rent)),
                         size = 0.5) +
  facet_wrap(~locality) +
  scale_x_yearqtr(format = "%Y") +
  scale_y_continuous(labels = label_dollar(),
                     n.breaks = 7) +
  scale_color_manual(values = submarkets) +
  labs(title = "Average asking rent by submarket",
       subtitle = "2010 Q1 to 2022 Q2",
       caption = "**Source:** CoStar Group, Inc.") +
  theme_pha() +
  theme(legend.position = "right",
        axis.text.x = element_text(angle = 90))

if (knitr::is_html_output()) {
  
  girafe(ggobj = sub_rent_plot,
         height_svg = 4) 
  
} else {
  
  sub_rent_plot
  
  }
  
```

### Rents by bedrooms

Rents in the region have risen the most among three-bedroom and two-bedroom apartments, reflecting continued demand for units that have actually *declined* in supply since 2016. In contrast, average rents for studio and one-bedroom apartments---which grew by more than 2,700 units since 2016---have increased less than \$100 over the last decade when adjusted for inflation.

```{r}
#| label: fig-avg-rent-br
#| fig.cap: "Average asking rent by bedroom"

# Aggregated average market asking rent per unit for Chesterfield, Hanover, Henrico, and Richmond City.

# CoStar Filters: Construction Status - Existing

# As of mid-JULY; inflation-adjusted

pha_rent_br <- read_csv("data/pha_rent_br.csv") |> 
  clean_names() |> 
  mutate(period = case_when(
    period == "2022 Q3 QTD" ~ "2022 Q3",
    TRUE ~ period)) |> 
  mutate(period = as.Date(as.yearqtr(period, format = "%Y Q%q"), frac = 1)) |>
  pivot_longer(cols = 2:5,
               names_to = "br",
               values_to = "rent") |> 
  mutate(br = case_when(
    br == "studio" ~ "Studio",
    br == "x1_bed" ~ "One bedroom",
    br == "x2_beds" ~ "Two bedrooms",
    br == "x3_beds" ~ "Three bedrooms")) |> 
  mutate(rent = parse_number(rent),
         br = fct_reorder(br, rent, .desc = TRUE)) |> 
  filter(between(period, as.Date("2010-03-31"), as.Date("2022-06-30"))) |> 
  mutate(period = as.yearqtr(period))
  
rent_br_plot <- ggplot(pha_rent_br,
       aes(x = period, y = rent, color = br)) +
  geom_line(size = 1) +
  geom_point_interactive(aes(data_id = rent, tooltip = label_dollar()(rent)),
                         size = 0.5) +
  scale_y_continuous(labels = label_dollar(),
                     n.breaks = 9) +
  scale_x_yearqtr(format = "%Y Q%q") +
  scale_color_pha() +
  labs(title = "Average asking rent by bedroom",
       subtitle = "2019 Q1 to 2022 Q2 (inflation-adjusted)",
       caption = "**Source:** CoStar Group, Inc.") +
  theme_pha() +
  theme(legend.position = "right")

if (knitr::is_html_output()) {
  
  girafe(ggobj = rent_br_plot,
         height_svg = 4) 
  
} else {
  
  rent_br_plot
  
  }

```

### Rents by age of units

Recently constructed rental housing (built in 2010 and after) leads average asking rents at \$1,614. As expected, rental costs correlate to the period in which they were built --- with older rental housing being less expensive. Pre-1980 rental housing is roughly \$400 cheaper than more recent rental housing.

In the last decade, more recent rental housing had steady and modest increases; only increasing \$80 from Q1 2012 to Q2 2022. But older rental housing had much more dramatic increases; increasing an average of \$257 in that same time period.

Rental housing built between 1980 and 2009 had especially steep increases during the height of the pandemic (Q1 2020 to Q3 2021). In this time, the average asking rent increased by over \$130, while rent increases for newer rental housing and pre-1980 housing increased by less than \$100. 

```{r}
#| label: fig-rents-built
#| fig.cap: "Average asking rent by age of unit"

# Aggregated average market asking rent per unit for Chesterfield, Hanover, Henrico, and Richmond City.

# CoStar Filters: Construction Status - Existing

# Not easily done in CoStar; first filter at max year 1979; then do 1980 to 1999, then 2000 to 2009, then 2010 and beyond; inflation-adjusted

rent_79 <- read_csv("data/pha_rent_79.csv") |> 
  clean_names() |> 
  mutate(built = "Pre-1980")

rent_99 <- read_csv("data/pha_rent_99.csv") |> 
  clean_names() |> 
  mutate(built = "1980 to 1999")

rent_09 <- read_csv("data/pha_rent_09.csv") |> 
  clean_names() |> 
  mutate(built = "2000 to 2009")

rent_10 <- read_csv("data/pha_rent_10.csv") |> 
  clean_names() |> 
  mutate(built = "2010 and later")

pha_built <- list(rent_79, rent_99, rent_09, rent_10)

pha_built_clean <- function(df){
  df |> 
    mutate(period = case_when(
    period == "2022 Q3 QTD" ~ "2022 Q3",
    TRUE ~ period
  )) |> 
  mutate(period = as.Date(as.yearqtr(period, format = "%Y Q%q"), frac = 1)) |>  
  mutate(rent = parse_number(current_search)) |> 
  select(period, built, rent)
}

pha_built_rent <- map_dfr(pha_built, pha_built_clean) |> 
  mutate(built = fct_reorder(built, rent, .desc = TRUE)) |> 
  filter(between(period, as.Date("2012-03-31"), as.Date("2022-06-30"))) |> 
  mutate(period = as.yearqtr(period))

built_rent_plot <- ggplot(pha_built_rent,
       aes(x = period, y = rent, color = built)) +
  geom_line(size = 1) +
  geom_point_interactive(aes(data_id = rent, tooltip = label_dollar()(rent)),
                         size = 0.5) +
  scale_y_continuous(labels = label_dollar(),
                     n.breaks = 9) +
  scale_x_yearqtr(format = "%Y Q%q") +
  scale_color_pha() +
  labs(title = "Average asking rent by age of unit",
       subtitle = "2012 Q1 to 2022 Q2 (inflation-adjusted)",
       caption = "**Source:** CoStar Group, Inc.") +
  theme_pha() +
  theme(legend.position = "right")

if (knitr::is_html_output()) {
  
  girafe(ggobj = built_rent_plot,
         height_svg = 4) 
  
} else {
  
  built_rent_plot
  
  }

```

## Rental vacancy

For much of the past two decades, vacancy rates have fluctuated seasonally as new people enter and leave the rental housing market. Across the region, submarkets have largely had vacancy rates below ten percent. In 2022, the regional average vacancy rate to-date was five percent. 

However, some submarkets in the region have lower than average vacancy rates; Hanover County (1 percent), Eastern Henrico (3 percent), Northside (3 percent), and East End (4 percent) have significantly lower vacancy rates.  

```{r}
#| label: fig-vacancy
#| fig.cap: "Vacancy rates by submarket"

rr_sub_vacancy <- rr_sub_rent |> 
  select(-rent) |> 
  mutate(vacancy_rate = parse_number(vacancy_rate)/100) |> 
  mutate(vacancy_rate = replace(vacancy_rate, vacancy_rate > 0.15, NA))

sub_vacancy_plot <- ggplot(rr_sub_vacancy,
       aes(x = period,
           y = vacancy_rate, 
           color = submarket)) +
  geom_line(size = 1) + 
  geom_point_interactive(aes(data_id = vacancy_rate,
                             tooltip = label_percent(accuracy = 0.1)(vacancy_rate)),
                         size = 0.5) +
  facet_wrap(~locality) +
  scale_x_yearqtr(format = "%Y") +
  scale_y_continuous(labels = label_percent(),
                     n.breaks = 7) +
  scale_color_manual(values = submarkets) +
  labs(title = "Vacancy rates by submarket",
       subtitle = "2010 Q1 to 2022 Q2",
       caption = "**Source:** CoStar Group, Inc.<br>**Note:** Outlier values above 15% omitted.") +
  theme_pha() +
  theme(legend.position = "right",
        axis.text.x = element_text(angle = 90)) +
  add_zero_line("y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = sub_vacancy_plot,
         height_svg = 4) 
  
} else {
  
  sub_vacancy_plot
  
  }

```
