==> part-2-3.qmd <==
# Housing assistance {#part-2-3}

```{r}
#| label: setup

library(tidyverse)
library(tidygeocoder)
library(leaflet)
library(mapview)
library(sf)
library(sp)
library(janitor)
library(viridis)
library(htmlwidgets)
library(scales)
library(lubridate)
library(tidytext)
library(kableExtra)
library(formattable)
library(readxl)
library(tigris)
library(ggspatial)
library(mapboxapi)
library(ggiraph)
library(ggtext)
library(hdatools)

```

```{r}
#| label: geo-setup
#| eval: false

sf_use_s2(TRUE)

options(tigris_use_cache = TRUE)

rr <- c("51085", "51760", "51075", "51145", "51087", "51127", "51036", "51041")

rr_boundary <- counties("VA") |> 
  filter(GEOID %in% rr)

rr_basemap <- layer_static_mapbox(
  location = rr_boundary,
  style_id = "light-v10",
  username = "mapbox"
  )

write_rds(rr_basemap, "data/rr_basemap.rds")

write_rds(rr_boundary, "data/rr_boundary.rds")

```

This chapter covers the range of housing assistance in the region supported by federal, state, and local programs.

## Affordable rental housing

```{r}
#| label: nhpd-data
#| eval: false

nhpd <- read_csv("data/nhpd_va_subsidies_7_2022.csv") |>
  clean_names()

nhpd_clean <- nhpd |>
  transform(fulladdress = paste(nhpd$street_address, nhpd$city, nhpd$state, nhpd$zip_code, sep= ", ")) |>
  select(nhpd_property_id, nhpd_subsidy_id, subsidy_name, subsidy_status,
         subsidy_subname, start_date, end_date, assisted_units, known_total_units,
         property_name, fulladdress, owner_name, owner_type, target_population, city)

localities <- c("CHARLES CITY", "CHESTER", "CHESTERFIELD", "GLEN ALLEN", "HENRICO", "HIGHLAND SPRINGS", "MIDLOTHIAN", "NORTH CHESTERFIELD", "Richmond", "RICHMOND", "RICHMOND CITY", "SANDSTON", "ASHLAND", "MECHANICSVILLE")

nhpd_clean_rr <- nhpd_clean |>
  subset(city %in% localities) |>
  mutate(across(c(6,7), ~ as.Date(.x, "%m/%d/%Y"))) |>
  mutate(property_name = case_when(
    str_detect(fulladdress, "^5409 HULL") ~ "STUDIOS AT SOUTH RICHMOND",
    fulladdress == "707 N HARRISON ST, RICHMOND, VA, 23220" ~ "THE NEW CLAY HOUSE",
    fulladdress == "1611 N 31ST ST, RICHMOND, VA, 23223" ~ "ARMSTRONG RENAISSANCE",
    fulladdress == "1517 JEFFERSON DAVIS HWY, RICHMOND, VA, 23224" ~ "JAMES RIVER APARTMENTS",
    fulladdress == "2023 W CARY ST, RICHMOND, VA, 23220" ~ "REDEVELOPMENT AND HOMEOWNERSHIP PROGRAM",
    fulladdress == "1910 4TH AVE, RICHMOND, VA, 23222" ~ "HOUSING REDEVEL-1910 4TH AV #1",
    TRUE ~ property_name))

```

```{r}
#| label: vh-lihtc
#| eval: false

rr_names <- c("Richmond City", "Hanover County", "Henrico County", "Chesterfield County",
              "Powhatan County", "Goochland County", "New Kent County", "Charles City County")

vh_lihtc_recent <- read_xlsx("data/vh-lihtc-2022-01.xlsx",
                      range = "A2:L1384",
                      trim_ws = TRUE) |>
  clean_names() |>
  filter(jurisdiction %in% rr_names) |>
  separate(cycle_name, into = c("cycle_year", "cycle_type"), sep = "\\s", extra = "merge") |>
  mutate(subsidy_name = "LIHTC",
         subsidy_status = "Active",
         start_date = lubridate::ymd(cycle_year, truncated = 2L)) |>
  filter(start_date > '2010-01-01') |>
  mutate(cycle_type = str_squish(cycle_type),
         cycle_type = str_replace(cycle_type, "Hsg", "Housing")) |>
  mutate(fulladdress = str_to_upper(paste0(street_address, ", ", city, ", VA", " ", zip))) |>
  mutate(subsidy_subname = case_when(
    str_detect(cycle_type, "9%") &  duplicated(vhda_number) ~ "9% Tax Credit and maybe 4%",
    str_detect(cycle_type, "4%") ~ "4% Tax Credit",
    TRUE ~ "TCEP")) |>
  select(nhpd_property_id = vhda_number, subsidy_name, subsidy_status,
         subsidy_subname, start_date, assisted_units = tax_credit_units,
         known_total_units = total_units, property_name, fulladdress,
         address_components.county = jurisdiction) |>
  filter(!property_name %in% c("Ashland Woods", "Audubon Village II", "Bellevue",
                               "Belle Summit", "Brook Creek Crossings",
                               "Cary West", "Eggleston Plaza", "Goodwyn at Union Hill",
                               "Greens at Virginia Center", "Hatcher Tobacco Flats",
                               "Highland Grove I", "Highland Grove II", "Iron Bridge Road",
                               "Jefferson Mews", "Kingsridge I", "Miller Lofts",
                               "New Clay House II", "North Oak", "Oakmeade",
                               "Somanath Seniors at Beckstoffers",
                               "Studios at South Richmond", "Townhomes at Warwick Place",
                               "Tuscany Townhomes")) |>
  mutate(nhpd_property_id = case_when(
    str_detect(property_name, "Armstrong Renaissance") ~ 1155578,
    property_name == "Audubon Village I" ~ 1095212,
    property_name == "Baker School" ~ 1154801,
    property_name == "Belt Atlantic" ~ 1095423,
    property_name == "Brookland Park" ~ 1094671,
    property_name == "Chickahominy Bluff" ~ 1094869,
    property_name == "Creekside Manor" ~ 1095227,
    property_name == "Harbour Square  CLP" ~ 1094615,
    property_name == "Henrico Arms" ~ 1106924,
    property_name == "Hope Village" ~ 1094544,
    property_name == "Newbridge Village" ~ 1094390,
    property_name == "Omni Park Place" ~ 1103488,
    property_name == "Place One" ~ 1094568,
    property_name == "Shockoe Hill I" ~ 1094312,
    property_name == "St. Luke" ~ 1021344,
    property_name == "Village South" ~ 1094687,
    property_name == "William Byrd" ~ 1094904,
    property_name == "Woodland Crossing" ~ 1107114,
    TRUE ~ nhpd_property_id
  ))

nhpd_clean_rr <- nhpd_clean_rr |>
  bind_rows(vh_lihtc_recent) |>
  group_by(nhpd_property_id, subsidy_status) |>
  add_count(nhpd_property_id, name = "subsidies") |>
  mutate(subsidy_type = case_when(
      subsidies > 1 ~ "Multiple",
      TRUE ~ subsidy_name
  )) |>
  ungroup() |>
  mutate(known_total_units = case_when(
    nhpd_property_id == 1155578 ~ 186,
    TRUE ~ known_total_units
  ))

write_rds(nhpd_clean_rr, "data/nhpd_clean_rr.rds")

```

```{r}
#| label: nhpd-geocode
#| eval: false

nhpd_coded <- nhpd_clean_rr |>
  geocode(address = fulladdress,
          method = 'geocodio',
          full_results = TRUE,
          unique_only = FALSE,
          lat = Latitude,
          long = Longitude)

nhpd_coded_rr <- nhpd_coded |>
  rename(locality = address_components.county...30) |>
  mutate(Latitude = case_when(
    property_name == "NEW MANCHESTER FLATS IX" ~ 37.517602,
    property_name == "KINGSRIDGE I" ~ 37.551666,
    property_name == "GRAND OAKS SENIORS" ~ 37.353614,
    property_name == "IRON BRIDGE ROAD" ~ 37.361812,
    TRUE ~ Latitude)) |>
  mutate(Longitude = case_when(
    property_name == "NEW MANCHESTER FLATS IX" ~ -77.432408,
    property_name == "KINGSRIDGE I" ~ -77.375593,
    property_name == "GRAND OAKS SENIORS" ~ -77.457398,
    property_name == "IRON BRIDGE ROAD" ~ -77.495881,
    TRUE ~ Longitude))

write_rds(nhpd_coded_rr, "data/nhpd_coded_rr.rds")

```

An array of federal housing assistance programs help low-income residents across the region with rental housing opportunities. Today, there are approximately 25,969 dedicated affordable rental homes found across 240 properties in the Richmond area. These include units both currently occupied and in development.

::: callout-tip

### Funding sources

In the Richmond region, many affordable rental properties also receive assistance from the Virginia Housing Trust Fund, as well as local sources such as CDBG and HOME grants. The City of Richmond also awards funding to affordable rental projects with its own trust fund. These awards often fill a financing "gap" and do not provide a majority of the total assistance for a development; as a result, they are not specifically reflected in the data.

:::

### Subsidy types

Over half (51 percent) of all affordable rental homes in the region rely *solely* on the LIHTC program. Another 31 percent have layered multiple subsidies together, reflecting the capital and funding requirements needed to develop new affordable housing.

The other significant source of dedicated affordable housing continues to be more than 3,600 Public Housing units managed by Richmond Redevelopment and Housing Authority.

::: callout-note
Descriptions of each rental assistance program are available on the [NHPD website](https://preservationdatabase.org/documentation/program-descriptions/).
:::

:::{.callout-important}

It is important to note that Section 8 described in this section is not the same as Section 8 Housing Choice Vouchers. Section 8 subsidies described in this section refer to HUD project-based rental assistance --- meaning that they are rental assistance that is tied to a specific development, whereas Section 8 Housing Choice Vouchers are tenant-based subsidies that a recipient can take wherever they can find housing.

::::


```{r}
#| label: fig-subsidy-overall
#| fig-cap: "Share of federally assisted units by subsidy"

nhpd_clean_rr <- read_rds("data/nhpd_clean_rr.rds")

nhpd_coded_rr <- read_rds("data/nhpd_coded_rr.rds")

subsidy <- nhpd_clean_rr |>
  filter(subsidy_status != "Inactive") |>
  group_by(subsidy_type, nhpd_property_id) |>
  summarise(units = max(known_total_units)) |>
  group_by(subsidy_type) |>
  summarise(units = sum(units)) |>
  mutate(percent = units/sum(units))

subsidy_plot <- ggplot(subsidy,
       aes(y = reorder(subsidy_type, units),
           x = percent,
           fill = subsidy_type,
           label = label_comma()(units),
           data_id = percent,
           tooltip = label_comma()(units))) +
  geom_col_interactive(size = 2) +
  theme_pha() +
  scale_fill_pha() +
  scale_color_pha() +
  scale_x_continuous(labels = percent_format(), n.breaks = 10, limits = c(0, 0.59)) +
  labs(title = "Share of federally assisted units by subsidy type",
       subtitle = "Data as of July 2022",
       caption = "**Source:** National Housing Preservation Database.") +
  flip_gridlines() +
  add_zero_line("x")

if (knitr::is_html_output()) {
  
  girafe(ggobj = subsidy_plot,
         height_svg = 4)
  
} else {
  
  subsidy_plot +
    geom_text(aes(x = percent + 0.03,
                  color = subsidy_type),
              size = 4) +
    scale_color_pha()
  
  }

```

### Layered subsidies

In an effort to maximize assistance, rental subsidy programs are often layered together into single projects. Among the 8,061 units with multiple subsidies, over half have either a 4% or 9% LIHTC tax credit---or both "twinned" together. Section 8 Housing Finance and Development Agency (HFDA) New Construction and Loan Management Set-Aside (LMSA) programs are also common types.

"Other" subsidies generally include HUD insurance programs and other, less common, Section 8 programs. Still, these minor assistance packages nevertheless provide helping subsidy to almost three-in-four units with multiple affordability contracts.

```{r}
#| label: tbl-multiple
#| tbl-cap: "Active subsidies in units with mutiple subsidies"

multiple <- nhpd_clean_rr |>
  filter(subsidy_status != "Inactive",
         subsidy_type == "Multiple") |>
  group_by(subsidy_name, subsidy_subname) |>
  summarise(units = sum(assisted_units)) |>
  ungroup() |>
  mutate(detail = case_when(
    subsidy_name == "LIHTC" ~ subsidy_subname,
    subsidy_name %in% c("HOME", "RHS 515", "Project Based Vouchers", "Section 202") ~ subsidy_name,
    units < 400 ~ " ",
    TRUE ~ subsidy_subname)) |>
  mutate(subsidy_name = case_when(
    subsidy_name %in% c("HOME", "RHS 515", "Project Based Vouchers", "Section 202") ~ "Other",
    TRUE ~ subsidy_name)) |>
  mutate(detail = case_when(
    subsidy_subname == "4% and 9% Tax Credit" ~ subsidy_subname,
    subsidy_name == "Section 8" & units < 400 ~ "Other",
    subsidy_subname == "207/ 223(F) PUR/ REFIN HSG" ~ "207/223(F)",
    subsidy_subname == "221(D)(4) MKT RATE MOD INC/ DISP FAMS" ~ "221(D)(4)",
    subsidy_name == "HUD Insured" & units < 600 ~ "Other",
    TRUE ~ detail
  )) |>
  mutate(subsidy_name_ext = paste(subsidy_name, detail),
         subsidy_name_ext = fct_lump_n(subsidy_name_ext, n = 4, w = units),
         detail = str_wrap(detail, width = 10))

multiple |>
  group_by(subsidy_name_ext) |>
  summarise(units = sum(units)) |>
  mutate(pct = units / 8061,
         pct = percent(pct, digits = 1),
         units = comma(units, digits = 0)) |>
  kable(align = "lrr",
        col.names = c("Detailed subsidy type", "Units with subsidy", "Percent of total")) |>
  kableExtra::kable_styling(bootstrap_options = c("condensed", "striped")) |>
  kableExtra::footnote(general = "Sources: National Housing Preservation Database and Virginia Housing.",
                       general_title = "")

```

::: callout-note

Table totals do not add to 100 percent because units are percent of all 8,061 *units* with multiple subsidies, not the total of all subsidies.

:::

```{r}
#| label: fig-multiple-detail
#| fig.cap: "Detailed subsidy type for units with multiple subsidies"
#| eval: false

multi_plot <- ggplot(multiple,
       aes(y = reorder_within(detail, units, subsidy_name),
           x = units,
           fill = subsidy_name,
           data_id = units,
           tooltip = label_comma()(units))) +
  geom_col(position = "dodge") +
  geom_col_interactive(position = "dodge", size = 2) +
  theme_pha() +
  scale_fill_pha() +
  facet_wrap(~subsidy_name, ncol = 2, nrow = 2, scales = "free_y") +
  scale_y_reordered() +
  scale_x_continuous(labels = label_comma()) +
  labs(title = "Detailed subsidy type for units with multiple subsidies",
       subtitle = "Data as of July 2022",
       caption = "**Source:** National Housing Preservation Database.") +
  theme(panel.grid.major.y = element_blank(),
        panel.grid.major.x = element_line(
        color = "#cbcdcc",
        size = 0.05
        ))

if (knitr::is_html_output()) {
  
  girafe(ggobj = multi_plot,
         height_svg = 4)
  
} else {
  
  multi_plot
  
  }

```

### Locations

The map below shows the locations of affordable rental properties in the Richmond region. Each color corresponds to the property's subsidy as recorded in the National Housing Preservation Database, or if multiple subsidies are currently active.

```{r}
#| label: nhpd-map
#| eval: false

nhpd_mapped_rr <- read_rds("data/nhpd_coded_rr.rds") |>
  st_as_sf(coords = c("Longitude", "Latitude"), crs=4326, remove = FALSE) |>
  filter(fulladdress != "PO BOX 8585, RICHMOND, VA, 23226")

write_rds(nhpd_mapped_rr, "data/nhpd_mapped_rr.rds")

```

```{r}
#| label: fig-nhpd-map
#| fig.cap: "Federally-assisted rental housing properties"

rr_basemap <- read_rds("data/rr_basemap.rds")

nhpd_map <- read_rds("data/nhpd_mapped_rr.rds") |>
  filter(subsidy_status != "Inactive") |>
  group_by(subsidy_type, nhpd_property_id) |>
  summarise(units = max(known_total_units))

subsidyPal <- colorFactor(viridis(8), nhpd_map$subsidy_name)

if (knitr::is_html_output()) {
  
  mapview(nhpd_map,
          zcol= "subsidy_type",
          legend = T,
          popup = F,
          burst = F,
          layer.name = "Subsidy type")
  
  } else {
    
  ggplot() +
    rr_basemap +
    geom_sf(nhpd_map, mapping = aes(color = subsidy_type), show.legend = "point") +
    labs(title = "Federally-assisted rental housing properties",
         color = "Subsidy type",
         caption = "Sources: National Housing Preservation Database and Virginia Housing.") +
    #fixed_plot_aspect(ratio = 0.8) +
    theme(panel.background = element_blank(),
          axis.title = element_blank(),
          axis.ticks = element_blank(),
          axis.text = element_blank(),
          legend.key = element_blank(),
          plot.caption = element_text(hjust = 0))
    
  }

```

Richmond continues to support the majority of affordable rentals in the region (about 60 percent---more than 15,200). While Chesterfield and Henrico counties both have similar amounts of LIHTC-only units, Henrico has nearly 2,900 additional units supported by multiple subsidies---generally combinations of LIHTC and a Section 8 program.

```{r}
#| label: fig-subsidy-location
#| fig-cap: "Federally assisted units by subsidy and locality"

subsidy_local <- nhpd_coded_rr |>
  select(nhpd_property_id, subsidy_type, known_total_units,
         subsidy_status, locality) |>
  filter(subsidy_status != "Inactive",
         locality != "Charles City County") |>
  group_by(locality, subsidy_type, nhpd_property_id) |>
  summarise(units = max(known_total_units)) |>
  group_by(locality, subsidy_type) |>
  summarise(units = sum(units)) |>
  filter(units > 1) |> 
  mutate(locality = str_wrap(locality, 12))

local_plot <- ggplot(subsidy_local,
       aes(x = reorder(locality, units),
           y = units,
           fill = reorder(subsidy_type, units),
           data_id = units,
           tooltip = label_comma()(units))) +
  geom_col() + 
  geom_col_interactive(size = 2) +
  theme_pha() +
  scale_fill_pha() +
  scale_y_continuous(labels = label_comma()) +
  labs(title = "Federally assisted units by subsidy type and locality",
       subtitle = "Data as of July 2022",
       fill = "Subsidy type",
       caption = "**Source:** National Housing Preservation Database.") +
  theme(legend.position = "right")

# ggplot(subsidy_local,
#        aes(y = reorder_within(subsidy_type, units, locality),
#            x = units,
#            fill = subsidy_type)) +
#   geom_col(position = "dodge") +
#   facet_wrap(~locality, ncol = 2, nrow = 2, scales = "free_y") +
#   scale_y_reordered() +
#   scale_x_continuous(labels = label_comma()) +
#   labs(title = "Federally assisted units by subsidy type and locality",
#        subtitle = "Data as of July 2022",
#        caption = "Source: National Housing Preservation Database.") +
#   theme(legend.position = "none",
#         panel.background = element_blank(),
#         axis.ticks = element_blank(),
#         axis.title = element_blank(),
#         panel.grid.major.x = element_line(color = "grey95",
#                                   size = 0.05))

if (knitr::is_html_output()) {
  
  girafe(ggobj = local_plot,
         height_svg = 4) 
  
} else {
  
  local_plot
  
  }

```

### Changes since 2020

Since January 2020, the region has seen 57 new rental subsidies added, which increased the number of active affordability contracts on units by 4,393. Over that same period, 17 subsidies ended, affecting 1,633 units. Some properties had multiple subsidies either added or expired. In all, there was a net addition of 2,760 rental affordability contracts.

```{r}
#| label: tbl-change-table
#| tbl-cap: "Added and removed affordable rental contracts since 2020"

change <- nhpd_coded_rr |>
  mutate(change_type = case_when(
    start_date > '2019-12-31' ~ "Added",
    end_date < '2022-09-01' & end_date > '2019-12-31' ~ "Removed",
    TRUE ~ "No change"))

change_table <- change |>
  group_by(change_type) |>
  summarise(subsidies = n(),
            properties = n_distinct(nhpd_property_id),
            units = sum(assisted_units)) |>
  filter(change_type != "No change") |>
  mutate(across(where(is.numeric), ~ ifelse(change_type == "Removed", (.x * -1), .x))) |>
  adorn_totals(name = "Net change")

change_table |>
  mutate(across(where(is.numeric), ~ comma(.x, digits = 0))) |>
  kable(align = "lrrr",
        col.names = c("", "Subsidies", "Properties affected", "Units included")) |>
  kableExtra::kable_styling(bootstrap_options = c("condensed", "striped")) |>
  kableExtra::row_spec(3, bold = TRUE) |>
  kableExtra::footnote(general = "Sources: National Housing Preservation Database and Virginia Housing.",
                       general_title = "")

```

New LIHTC units (net 2,129) comprised the majority of added affordable rentals, followed by Section 8 contracts (net 1,132). Net losses of affordable rental contracts occurred in projects supported by HOME funding and HUD insurance.

```{r}
#| label: fig-change-fig
#| fig.cap: "Additions and removals of subsidized rental units"

change_subsidy <- change |>
  group_by(locality, subsidy_name, change_type) |>
  summarise(units = sum(assisted_units)) |>
  filter(change_type != "No change") |>
  mutate(units = if_else(change_type == "Removed", units*-1, units)) |>
  pivot_wider(
    names_from = change_type,
    values_from = units) |>
  mutate(across(where(is.numeric), ~ replace_na(.x, 0)),
         `Net change` = Removed + Added) |>
  pivot_longer(
    cols = 3:5,
    names_to = "change_type",
    values_to = "units") |>
  mutate(subsidy_name = fct_reorder(subsidy_name, units, .desc = TRUE),
         change_type = fct_relevel(change_type, "Added", "Removed", "Net change"),
         subsidy_name = str_wrap(subsidy_name, width = 10))

change_region <- change_subsidy |>
  filter(change_type == "Net change") |>
  group_by(subsidy_name) |>
  summarise(units = sum(units))

chg_plot <- ggplot(change_region,
       aes(x = reorder(subsidy_name, units),
           y = units,
           fill = subsidy_name,
           label = label_number(big.mark = ",", style_positive = "plus",
                                style_negative = "minus")(units),
           data_id = units,
           tooltip = label_number(big.mark = ",", style_positive = "plus",
                                style_negative = "minus")(units))) +
  geom_col_interactive(size = 2) +
  theme_pha() +
  scale_fill_pha() +
  scale_y_continuous(labels = label_comma(),
                     limits = c(-1750, 3100),
                     n.breaks = 10) +
  labs(title = "Net change in subsidized rental unit contracts",
       subtitle = "January 2020 through August 2022",
       caption = "**Source:** National Housing Preservation Database and Virginia Housing.") +
  scale_y_continuous(labels = label_number(big.mark = ",", style_positive = "plus", style_negative = "minus")) +
  add_zero_line("y")


if (knitr::is_html_output()) {
  
  girafe(ggobj = chg_plot,
         height_svg = 4) 
  
} else {
  
  chg_plot +
    geom_text(aes(y = units + sign(units)/0.003,
                  color = subsidy_name)) +
    scale_color_pha()
  
  }

```

While LIHTC additions drove new affordable supply in Richmond and Chesterfield, new (or renewed) Section 8 contracts covered more than 400 units in Henrico. Over 500 units in Richmond and Henrico saw HUD insurance contracts expire; however, many of these are in projects with another form of rental assistance that has not expired.

```{r}
#| label: fig-change-locality
#| fig.cap: "Net change in subsidized rental unit contracts by locality"

change_subsidy <- change |>
  group_by(locality, subsidy_name, change_type) |>
  summarise(units = sum(assisted_units)) |>
  filter(change_type != "No change") |>
  mutate(units = if_else(change_type == "Removed", units*-1, units)) |>
  pivot_wider(
    names_from = change_type,
    values_from = units) |>
  mutate(across(where(is.numeric), ~ replace_na(.x, 0)),
         `Net change` = Removed + Added) |>
  pivot_longer(
    cols = 3:5,
    names_to = "change_type",
    values_to = "units") |>
  mutate(subsidy_name = fct_reorder(subsidy_name, units, .desc = TRUE),
         change_type = fct_relevel(change_type, "Added", "Removed", "Net change"),
         subsidy_name = str_wrap(subsidy_name, width = 10))

change_locality <- change_subsidy |>
  filter(change_type == "Net change",
         locality != "Charles City County")

lchg_plot <- ggplot(change_locality,
       aes(x = reorder(subsidy_name, units),
           y = units,
           fill = subsidy_name,
           data_id = units,
           tooltip = label_number(big.mark = ",", style_positive = "plus",
                                  style_negative = "minus")(units))) +
  geom_col() + 
  geom_col_interactive(size = 1) +
  theme_pha() +
  scale_fill_pha() +
  facet_wrap(~locality) +
  scale_y_continuous(labels = label_comma()) +
  labs(title = "Net change in subsidized rental unit contracts by locality",
       subtitle = "January 2020 through August 2022",
       caption = "**Sources:** National Housing Preservation Database and Virginia Housing.") +
  scale_y_continuous(labels = label_number(big.mark = ",", style_positive = "plus", style_negative = "minus")) +
  add_zero_line("y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = lchg_plot,
         height_svg = 4) 
  
} else {
  
  lchg_plot
  
  }

```

### LIHTC preservation

LIHTC properties have a 30 year commitment to affordability, but only a 15 year compliance period, wherein property owners can increase rents. Nonprofit developers will often seek new allocation of tax credits before their commitment period ends, but there is often little incentive for for-profit developers to maintain affordability restrictions past the compliance period.

By 2040, a large portion of active LIHTC units will be outside the 30 year commitment period --- even far more will be outside the 15 year compliance period. Just over 13,000 LIHTC units will be beyond the 30 year commitment period by 2040, which accounts for well over half of all active LIHTC units as of early 2022.

```{r}
#| label: fig-lihtc
#| fig-cap: "Percent of active LIHTC units by end of commitment period"

lihtc <- nhpd_coded_rr |>
  filter(subsidy_name == "LIHTC") |>
  select(nhpd_property_id, property_name, subsidy_name, subsidy_type, subsidy_status, owner_type, start_date, end_date, known_total_units, assisted_units, locality) |>
  mutate(end_date = case_when(
    is.na(end_date) ~ as.Date(start_date %m+% years(30)),
    TRUE ~ end_date)) |>
  filter(subsidy_status == "Active") |>
  group_by(nhpd_property_id) |>
  summarise(start_date = min(start_date),
            end_date = max(end_date),
            units = max(known_total_units)) |>
  ungroup() |>
  group_by(end_year = ceiling_date(end_date, "5 years")) |>
  summarise(units = sum(units)) |>
  mutate(pct = units/sum(units),
         sum = cumsum(units),
         pct_cum = sum/sum(units),
         date = as.character(year(end_year)))

lihtc_plot <- ggplot(lihtc,
       aes(x = date,
           y = pct,
           fill = pct,
           data_id = pct,
           tooltip = percent_format(accuracy = 0.1)(pct))) +
  geom_col_interactive(size = 2) +
  scale_fill_gradientn(colours = c("#CAE3C2", "#74BA91", "#519B8D", "#2B4258")) +
  theme_pha() +
  scale_y_continuous(labels = label_percent(),
                     n.breaks = 6) +
  labs(title = "Percent of active LIHTC units by end of commitment period",
       subtitle = "Included properties with multiple subsidies",
       caption = "**Sources:** National Housing Preservation Database and Virginia Housing.") +
  add_zero_line("y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = lihtc_plot,
         height_svg = 4) 
  
} else {
  
  lihtc_plot
  
  }



```

### Public Housing

The redevelopment of public housing in the City of Richmond has begun to take shape at the first of the "Big Six" public housing courts --- [**Creighton Court**](https://www.rrha.com/redevelopment/creighton/). This public housing property located in Richmond's far East End consisted of 504 public housing units.

As part of their first phase of redevelopment, RRHA has begun demolition of Creighton Court, with plans to develop roughly 700 units of mixed-income housing. Construction on Phase 1 is expected to begin in Winter 2022 with the entire redevelopment process expected to last ten years.

RRHA's next focus area will be [**Gilpin Court**](https://www.rrha.com/redevelopment/jackson-ward/), north of Jackson Ward, where about 780 public housing units reside. In November 2021, RRHA and the City of Richmond were awarded a Choice Neighborhoods Planning Grant by the U.S. Department of Housing and Urban Development for \$450,000.

This planning grant is being utilized to facilitate the community planning process around not only Gilpin Court, but the Jackson Ward community --- including strategies to undo the negative impacts of interstate development on the historically Black communities of Jackson Ward and North Jackson Ward.

| Public housing community | Original units | Replacement units | Net change |
|--------------------------|---------------:|------------------:|-----------:|
| Creighton Court          |         504    |           700     |    +196    |
| Gilpin Court             |         780    |           TBD     |     TBD    |

: Net change in units for public housing redevelopment

::: callout-tip

#### Funding for Public Housing redevelopment

Federal housing policy has guided public housing authorities to use newer funding streams to redevelop older public housing communities. For these efforts, RRHA and its partners will use LIHTC, Tenant Protection Vouchers, Project-Based Vouchers, and other federal, state, and local sources.

:::

## Rental assistance

### Housing Choice Vouchers

Section 8 Housing Choice Vouchers (HCV) are tenant-based rental assistance that allows recipients to find housing on the open market. This provides household with greater choice in where they want to live, but historically many HCV recipients have faced discrimination from landlords unwilling to accept HCV.

This changed significantly in early 2020, when the Virginia General Assembly passed [new fair housing legislation](https://homeofva.org/get-help/fair-housing/source-of-funds/) that made it illegal to discriminate based on source of income --- defined as:

> *"any source that lawfully provides funds to or on behalf of a renter or buyer of housing, including any assistance, benefit, or subsidy program, whether such program is administered by a governmental or nongovernmental entity."*

HCV utilization across the region is concentrated in the East End and Southside of Richmond, but can also be found throughout the counties, as well as the Town of Ashland. Higher HCV utilization (above 20 percent) is seen in areas near Fulton Hill, Oakwood, Manchester, and Bellwood in Chesterfield County.


```{r}
#| label: hcv-setup
#| eval: false

# Download HCV geojson from HUD - 7.28.22

# HCV as a Percent of Renter Occupied Housing Units

hcv <- st_read("data/va_hcv_tract.geojson",
               quiet = TRUE) |>
  clean_names() |>
  subset(county %in% pha) |>
  filter(!is.na(hcv_public_pct))

write_rds(hcv, "data/hcv_tracts.rds")

```

```{r}
#| label: fig-hcv
#| fig.cap: "Percent of renters with Housing Choice Vouchers by tract"

hcv <- read_rds("data/hcv_tracts.rds")

if (knitr::is_html_output()) {
  
  mapview(hcv,
          zcol= "hcv_public_pct",
          legend = T,
          layer.name = "Percent renters with HCVs")
  
  } else {
    
  ggplot() +
    rr_basemap +
    layer_spatial(hcv, aes(fill = hcv_public_pct)) +
    scale_fill_gradient2(low = "#edf8b1", high = "#253494", mid = "#41b6c4",
                         midpoint = 18,
                         labels = label_percent(scale = 1)) +
    labs(title = "Percent of renters\nwith Housing Choice Vouchers by tract",
         fill = "Percent",
         caption = "Source: U.S. Department of Housing and Urban Development.") +
    #fixed_plot_aspect(ratio = 0.8) +
    theme(panel.background = element_blank(),
          axis.title = element_blank(),
          axis.ticks = element_blank(),
          axis.text = element_blank(),
          legend.key.height = unit(0.1, "npc"),
          plot.caption = element_text(hjust = 0))
    
  }

```

### Rent relief and mortgage relief

In response to the COVID-19 pandemic's impact on renters across the nation, Congress created a \$25 billion Emergency Rental Assistance (ERA) program that was funded through the CARES Act in 2021. The program was implemented through the U.S. Treasury Department and resulted in a total of $1 billion being allocated to the Commonwealth of Virginia and eligible local governments.

With this funding, the Virginia Department of Housing and Community Development (DHCD) established the Virginia Rent Relief Program (RRP), while Chesterfield County elected to administer their own rental assistance program through local nonprofit, Area Congregations Together in Service (ACTS).

Through DHCD, a total of 32,029 payments were made to households across Richmond, Henrico, and Hanover. Both Richmond and Henrico saw increasing households receiving rental assistance with slight dips during the early part of 2022. Few households in Hanover County sought rental relief from DHCD.

::: callout-warning

#### All data not available

Data for Chesterfield County's separately-administered rent relief program was not available. 

:::

```{r}
#| label: fig-rrp
#| fig.cap: "Rent relief payments by locality"

# Email gabriella.vazquez@dhcd.virginia.gov for data on RRP - breakdown of households helped in Hanover, Henrico, and Richmond --- consider including other localities

# Email Jessica about Chesterfield rent relief program.

rrp <- read_csv("data/rrp_data.csv")|>
  mutate(avgpayment = spent/payments) |>
  filter(locality == "Hanover County" | locality == "Henrico County" | locality == "Richmond") |>
  mutate(date = mdy(date))

rrp_plot <- ggplot(rrp,
       aes(x = date,
           y = payments,
           fill = locality,
           data_id = payments,
           tooltip = label_comma()(payments))) +
  geom_col() +
  geom_col_interactive(size = 1) +
  theme_pha() +
  scale_fill_manual(values = c("#a6cccc", "#f39152", "#be451c")) +
  facet_wrap(~locality) +
  labs(title = "Rent relief payments by locality",
       subtitle = "Through June 2022",
       caption = "**Source:** Virginia Department of Housing and Community Development") +
  scale_x_date(date_labels = "%b %y",
               expand = c(0,0)) +
  scale_y_continuous(labels = label_comma()) +
  theme(axis.text.x = element_text(angle = 90, vjust = 1, hjust = 1)) +
  add_zero_line("y")

# https://crsreports.congress.gov/product/pdf/R/R46688
  
if (knitr::is_html_output()) {
  
  girafe(ggobj = rrp_plot,
         height_svg = 4) 
  
} else {
  
  rrp_plot
  
  }

```

Average payments per household across the three localities was well above \$4,000 for most of the program's duration, with the highest payments being made in Hanover County.

```{r}
#| label: fig-rrp-payment
#| fig.cap: "Average assistance by month and locality"

rrpay_plot <- ggplot(rrp,
       aes(x = date,
           y = avgpayment,
           fill = locality,
           data_id = avgpayment,
           tooltip = dollar_format()(avgpayment))) +
  geom_col() +
  geom_col_interactive(size = 1) +
  theme_pha() +
  scale_fill_manual(values = c("#a6cccc", "#f39152", "#be451c")) +
  facet_wrap(~locality) +
  labs(title = "Average assistance by month and locality",
       subtitle = "Through June 2022",
       caption = "**Source:** Virginia Department of Housing and Community Development") +
  scale_x_date(date_labels = "%b %y",
               expand = c(0,0)) +
  scale_y_continuous(labels = label_dollar()) +
  theme(axis.text.x = element_text(angle = 90, vjust = 1, hjust = 1)) +
  add_zero_line("y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = rrpay_plot,
         height_svg = 4) 
  
} else {
  
  rrpay_plot
  
  }


```

## Affordable homeownership

Since 2018, nonprofit developers in the region have averaged about 53 affordable homes sold to low-income buyers a year. Most of this production is attributable to Southside Community Development and project:HOMES.


```{r}
#| label: fig-pha-homeownership
#| fig-cap: "Richmond region nonprofit homeownership production"

pha_ho <- read_csv("data/pha_homeownership.csv") |>
  clean_names() |>
  select(1:7)

ho_production <- pha_ho |>
  filter(data_type == "Number of homes completed and sold in calendar year") |>
  pivot_longer(3:7,
               names_to = "year",
               values_to = "homes") |>
  mutate(year = parse_number(year))

avg_ho <- ho_production %>% 
  group_by(year) %>% 
  summarise(total = sum(homes))

# mean(avg_ho$total)

# MWCLT, BHC, project:HOMES, Southside, Habitat - WAITING ON DATA

# Ask for number of units produced each year from 2019 to now

# Average home price each year?

ho_plot <- ggplot(ho_production,
       aes(x = year,
           y = homes,
           fill = reorder(organization, homes),
           data_id = homes,
           tooltip = homes)) +
  geom_col() +
  geom_col_interactive(size = 1) +
  theme_pha() +
  scale_fill_pha() +
  labs(title = "Richmond region nonprofit homeownership production",
       subtitle = "Numbers for 2022 are current through August 2022",
       caption = "**Source:** Survey of nonprofit organizations.") +
  theme(legend.position = "bottom") +
  guides(fill = guide_legend(nrow = 2))

if (knitr::is_html_output()) {
  
  girafe(ggobj = ho_plot,
         height_svg = 4) 
  
} else {
  
  ho_plot
  
  }


```
