==> part-2-4.qmd <==
# Assessment of naturally-occurring affordable housing {#part-2-4}

```{r}
#| label: setup

library(tidyverse)
library(leaflet)
library(mapview)
library(htmlwidgets)
library(janitor)
library(sf)
library(sp)
library(scales)
library(zoo)
library(lubridate)
library(patchwork)
library(ggtext)
library(ggiraph)
library(cowplot)
library(hdatools)

# Utilize the same filters as the Chesterfield County Rental Market Analysis. I've changed the star rating to two or less, which coincides with CoStar's analysis.

# CoStar Filters for Properties: Multi-Family, Localities: Chesterfield, Hanover, Henrico,Richmond City; Construction status: Existing; Stars: 2 or less; Rent type: Market; Max Year Built: 1999; Building class: B, C, and Unclassified

# This returns a total of 479 properties as of 8.17.22

# There are about 80 properties that do not have a listed style. Some may be manufactured home parks. Consider removing property_name == Sedgefield, Holiday Mobile Home Park

# Append data from MHCCV report for manufactured home communities.

pha <- c("51085", "51760", "51087", "51041")

pha_names <- c("Richmond City", "Hanover", "Henrico", "Chesterfield")

```
This chapter covers trends in naturally-occurring affordable housing (NOAH), or market affordable housing.

## Naturally-occurring affordable housing

Not all affordable housing is supported by public subsidy. In fact, a large share of affordable housing is privately-owned and receives no government assistance. Widely referred to as [naturally occurring affordable housing (NOAH)](https://www.mckinsey.com/industries/public-and-social-sector/our-insights/preserving-the-largest-and-most-at-risk-supply-of-affordable-housing), or market-affordable housing, these properties are a growing concern for communities facing housing affordability challenges.

:::{.callout-note}

## Is it really "naturally-occurring"?

Many experts consider NOAH to be a [misnomer](https://cityobservatory.org/the-myth-of-naturally-occurring-affordable-housing/) because there is nothing "natural" about the affordability of these properties. Some prefer ["market-rate affordable"](https://www.allianceforhousingsolutions.org/market-rate-affordable-housing). But regardless of the choice of term, this type of housing plays a pivotal, albeit precarious, role in providing localities with a significant amount of affordable housing without government resources.

:::
The preservation of NOAH properties has been a growing strategy to support affordable housing in communities as the ability to quickly develop new units has been stifled by labor shortages, rising land prices, and supply chain issues. NOAH is at great risk of being lost because it often requires greater investment to maintain and more likely to be redeveloped---in turn contributing to a loss of affordable housing units.

::: callout-note

### Defining NOAH

For analysis purposes, we define NOAH properties as:

* Existing multifamily properties with active leases;
* Classified as Class B, C, or unclassified;
* No public subsidy, rent caps, or other income-based restrictions;
* *CoStar Building Rating* of two or fewer stars out of five;^[Per CoStar, this is rating for the building relative to other buildings of the same type throughout the country.] and
* Built before 2000

:::

### Locations

Based on this criteria, NOAH properties are located all across the region, especially in the City of Richmond and along the borders of the city and counties, where older multifamily properties exist.

```{r}
#| label: fig-noah-map
#| fig.cap: "NOAH properties in Richmond region"

rr_basemap <- read_rds("data/rr_basemap.rds")
  
noah <- read_csv("data/pha_noah_81722.csv") |> 
  clean_names() |> 
  st_as_sf(coords = c("longitude", "latitude"), crs=4326, remove=FALSE) |> 
  filter(!property_name %in% c("Sedgefield", "Holiday Mobile Home Park")) |> 
  mutate(property_name = fct_reorder(property_name, avg_asking_unit, .desc = TRUE))

if (knitr::is_html_output()) {
  
  mapview(noah,
    zcol = "avg_asking_unit",
    legend = T,
    popup = F,
    burst = F,
    layer.name = "Average asking rent ($)")
  
  } else {
  
  noah_dropna_rent <- noah |> 
    filter(!is.na(avg_asking_unit))

  ggplot() +
    rr_basemap +
    geom_sf(noah_dropna_rent, mapping = aes(color = avg_asking_unit)) +
    scale_color_continuous(type = "viridis", labels = label_dollar(), trans = "reverse") +
    scale_x_continuous(expand = expansion(mult = 0),
                       limits = c(-8660000, -8585000)) +
    scale_y_continuous(expand = expansion(mult = 0),
                       limits = c(4466959, 4548000)) +
    labs(title = "NOAH properties in Richmond region",
         subtitle = "Rents as of August 2022",
         color = "Average asking rent",
         caption = "**Source:** CoStar Group, Inc.") +
    theme_pha() +
    theme(panel.grid = element_blank(),
          axis.title = element_blank(),
          axis.text = element_blank(),
          legend.position = "right",
          legend.title = element_text(),
          legend.key = element_blank())
    
  }

```

The large share of NOAH is located in the City of Richmond (11,253 units or 45 percent of all NOAH units). This is no doubt due to the large amount of older multifamily properties within city limits --- most of which is located in smaller buildings. Henrico County has the second largest share of NOAH in the region with 8,983 units, followed by Chesterfield County at 3,667 and Hanover County below 500 units.

:::{.callout-warning}

### Additional context for NOAH

It is important to consider that the City of Richmond contains much of the region's older housing stock. This includes many older apartment buildings that are anywhere from two to twenty units in size. While they may fall under the NOAH classification set out in this analysis, their rents may actually be higher due to the overwhelming demand for rental housing among young professionals and students in the city.

Comparably cheaper than new multifamily properties that offer amenities such as pools and fitness centers, these properties are still able to command such high rents due to their proximity to Virginia Commonwealth University and Virginia Union University, as well as other urban amenities like restaurants, bars, and retail.

:::

In total, there are nearly 25,000 NOAH units across 194 properties in the region. This makes up a significant amount of rental inventory, as well as a valuable source of unsubsidized affordable housing. The aging of these properties, as well as the increasing demand for rental housing, puts significant pressure on these properties and their owners. For smaller landlords, the cost to renovate can be far too great --- leading to worsening deferred maintenance and pressure to sell, which both can negatively impact renters.

```{r}
#| label: fig-noah-cnt
#| fig.cap: "NOAH units by locality"

# Total units is 24,395

noah_cnt <- noah |> 
  st_drop_geometry() |> 
  group_by(county_name) |> 
  summarise(n = sum(number_of_units))

noah_cnt_plot <- ggplot(noah_cnt,
       aes(x = reorder(county_name, n),
           y = n,
           fill = county_name,
           label = label_comma()(n),
           data_id = n,
           tooltip = label_comma()(n))) +
  theme_pha() +
  geom_col() +
  geom_col_interactive(size = 2) +
  scale_color_manual(values = c("#5bab8e", "#a6cccc", "#f39152", "#be451c")) +
  scale_fill_manual(values = c("#5bab8e", "#a6cccc", "#f39152", "#be451c")) +
  scale_y_continuous(labels = label_comma()) +
  labs(title= "NOAH units by locality",
       subtitle = "Data through August 2022",
       caption = "**Source:** CoStar Group, Inc.") +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = noah_cnt_plot,
         height_svg = 4)
  
} else {
  
  noah_cnt_plot +
    geom_text(aes(y = n + 500,
                  color = county_name))
  
  }

```

### Building style

CoStar places buildings into style categories based on the following definitions:

| Style     | Stories      | Buildings           |
|-----------|--------------|---------------------|
| Garden    | 1-3 Stories  | 4 or more buildings |
| Low-Rise  | 1-3 Stories  | 1-3 buildings       |
| Mid-Rise  | 4-14 Stories | 1 or more buildings |
| High Rise | 15+ Stories  | 1 or more buildings |

While not traditional multifamily properties, CoStar does track some single-family and townhome rentals that are included in this data.

Roughly four out of five NOAH units are part of garden style properties, which are clusters of smaller one- to three-story buildings. Another 3,200 units in low-rise properties, which have similar but fewer buildings per community, round out nearly all of the NOAH supply in the region.

However, garden style properties have more than 210 units on average, while low-rise properties have just 12. As a result, there more than twice as many low-rise properties (273) than garden properties (95).

```{r}
#| label: noah-style

noah_style <- noah |> 
  st_drop_geometry() |> 
  mutate(style = replace_na(style, "Not listed")) |> 
  group_by(style) |> 
  summarise(units = sum(number_of_units),
            props = n()) |> 
  mutate(avg = units/props)

style_units <- ggplot(noah_style,
    aes(x = reorder(style, units, decreasing = TRUE),
        y = units,
        fill = style,
        data_id = units,
        tooltip = label_comma()(units))) +
  geom_col() +
  geom_col_interactive(size = 2) +
  theme_pha() + 
  scale_fill_pha() +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 16)) +
  scale_y_continuous(labels = label_comma()) +
  labs(subtitle = "Number of units",
       caption = "**Source:** CoStar Group, Inc.") +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0))

style_bldgs <- ggplot(noah_style,
    aes(x = reorder(style, units, decreasing = TRUE),
        y = props,
        fill = style,
        data_id = props,
        tooltip = label_comma()(props))) +
  geom_col() + 
  geom_col_interactive(size = 2) +
  theme_pha() + 
  scale_fill_pha() +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 16)) +
  scale_y_continuous(labels = label_comma()) +
  labs(subtitle = "Number of properties",
       caption = "**Source:** CoStar Group, Inc.")  +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0))

```

:::: {.content-visible when-format="html"}

::: {.panel-tabset}

#### Units

```{r}
#| label: fig-noah-style-units
#| fig.cap: "NOAH units by building style"

girafe(ggobj = style_units,
         height_svg = 4)

```

#### Properties

```{r}
#| label: fig-noah-style-bldgs
#| fig.cap: "NOAH properties by building style"

girafe(ggobj = style_bldgs,
         height_svg = 4)

```

:::

::::

::: {.content-visible when-format="pdf"}

```{r}
#| label: fig-noah-style
#| fig.cap: "NOAH units and properties by building style"

style_units <- style_units +
  theme(plot.caption = element_blank())

style_bldgs <- style_bldgs +
  theme(plot.caption = element_blank())

style_units + style_bldgs +
  plot_annotation(
    title = "NOAH units and properties by building style",
    caption = "**Source:** CoStar Group, Inc.",
    theme = theme_pha()
  )

```

:::

The map below shows NOAH properties by building style across the region. Most of the low-rise properties are within Richmond, reflecting the early 20th century small-scale apartment buildings found across many historic neighborhoods in the city---especially the Fan and Museum District.

Garden style properties, on the other hand, are more commonly found in the counties' inner suburbs, and reflect development trends prevalent during those areas' growth in the mid 20th century.

```{r}
#| label: fig-noah-style-map
#| fig.cap: "NOAH properties by building style"

if (knitr::is_html_output()) {
  
  mapview(
    noah,
    zcol = "style",
    legend = T,
    popup = F,
    burst = F,
    layer.name = "Building style"
  )

  } else {
  
  noah_dropna_style <- noah |> 
  filter(!is.na(style))

  ggplot() +
    rr_basemap +
    geom_sf(noah_dropna_style,
            mapping = aes(color = style),
            alpha = 0.5,
            show.legend = "point") +
    scale_color_pha() +
    scale_x_continuous(expand = expansion(mult = 0),
                       limits = c(-8660000, -8585000)) +
    scale_y_continuous(expand = expansion(mult = 0),
                       limits = c(4466959, 4548000)) +
    labs(title = "Building style of NOAH properties",
         subtitle = "Data as of August 2022",
         color = "Building style",
         caption = "**Source:** CoStar Group, Inc.") +
    theme_pha() +
    theme(panel.grid = element_blank(),
          axis.title = element_blank(),
          axis.text = element_blank(),
          legend.position = "right",
          legend.title = element_text(),
          legend.key = element_blank())
    
  }

```

### Age

With the exception of Richmond's low-rise apartments from the 1910s to 1930s, most of the region's NOAH units were built between 1960 and 1980. CoStar also tracks property renovations, which began for NOAH properties in the 2000s, especially closer to 2020 in Richmond and Chesterfield.

::: callout-tip

#### Substantial versus minor renovations

CoStar only marks a property as renovated if it:

"...*has been completely restored so that the existing space becomes 'new' space again. [...] Minor renovations, such as the improvement of a building's lobby or exterior are not considered full building renovations.*"

:::

```{r}
#| label: fig-noah-age
#| fig.cap: "NOAH properties by year built"

noah_age <- noah |> 
  st_drop_geometry() |> 
  mutate(year = case_when(
    is.na(year_renovated) ~ year_built,
    TRUE ~ year_renovated
  )) |> 
  mutate(status = case_when(
    is.na(year_renovated) ~ "Original",
    TRUE ~ "Renovated",
  )) |> 
  filter(year > 1900) |> 
  group_by(year, county_name, status) |> 
  summarise(number_of_units = sum(number_of_units))

noah_age_plot <- ggplot(noah_age,
       aes(x = year,
           y = number_of_units,
           fill = status,
           data_id = number_of_units,
           tooltip = label_comma()(number_of_units))) +
  geom_col() +
  geom_col_interactive(size = 2) +
  theme_pha() +
  scale_fill_pha() +
  facet_wrap(~county_name, nrow = 4, scales = "free_x") +
  scale_x_continuous(limits = c(1900, 2022),
                     breaks = c(1900, 1925, 1950, 1975, 2000, 2010, 2020)) +
  scale_y_continuous(labels = label_comma(),
                     breaks = c(500,1000),
                     expand = expansion(mult = 0)) +
  labs(title= "NOAH units by year built or renovated",
       subtitle = "Data through August 2022",
       caption = "**Source:** CoStar Group, Inc.") +
  add_zero_line(axis = "y") +
  theme(legend.position = "right",
        axis.ticks.x = element_line(size = 0.5, color = "#4b4f50"),
        plot.subtitle = element_markdown(margin = margin(t = 5, b = 0, unit = "pt")))

if (knitr::is_html_output()) {
  
  girafe(ggobj = noah_age_plot,
         height_svg = 4)
  
} else {
  
  noah_age_plot
  
  }

```

### Rents

NOAH properties have not been immune from the rapid rises in rent. As of Q3 2022, the average asking rent for NOAH properties was \$1,173. This is about \$200 less than the average asking rent across all rental properties. Although this seems like a small difference, an extra \$200 a month means more money saved for childcare or transportation costs.

From the beginning of the pandemic to Q3 2022, NOAH rent has increased by \$104 --- a 10 percent increase. 

```{r}
#| label: fig-noah-rents
#| fig-cap: "Average asking rent for market-affordable multifamily"

# Pulled at end of July; hesitant to re-pull because the data keeps changing on CoStar's side. 

noah_rent <- read_csv("data/pha_noah_rent_81722.csv") |> 
  mutate(period = as.Date(as.yearqtr(period, format = "%Y Q%q"), frac = 1)) |> 
  pivot_longer(cols = 2:3,
               names_to = "type",
               values_to = "rent") |> 
  mutate(type = case_when(
    type == "inf_adj" ~ "Inflation-adjusted",
    type == "rent" ~ "Nominal"
  )) |> 
  mutate(rent = parse_number(rent)) |> 
  filter(between(period, as.Date("2014-03-31"), as.Date("2022-09-30"))) |> 
  mutate(period = as.yearqtr(period))

noah_rent_plot <- ggplot(noah_rent,
       aes(x = period, 
           y = rent, 
           color = type)) +
  geom_line(size = 1) +
  geom_point_interactive(
    aes(data_id = rent,
        tooltip = label_dollar()(rent)),
    size = 0.5) + 
  theme_pha() +
  scale_color_pha() + 
  scale_y_continuous(labels = label_dollar(),
                     n.breaks = 9) +
  scale_x_yearqtr(format = "%Y Q%q") +
  labs(title = "Average asking rent for market-affordable multifamily",
       subtitle = "2014 Q1 to 2022 Q3 (to-date)",
       color = "Rent value",
       caption = "**Source:** CoStar Group, Inc.") +
  theme(legend.position = "top",
        plot.subtitle = element_markdown(margin = margin(t = 5, b = 0, unit = "pt")))

if (knitr::is_html_output()) {
  
  girafe(ggobj = noah_rent_plot,
         height_svg = 4)
  
} else {
  
  noah_rent_plot
  
  }

```

### Sales

```{r}
#| label: noah-sales-date
#| eval: false

# Search CoStar SALES: 
# NOAH --- Location: Richmond city, Henrico County, Chesterfield County, Hanover County
# Filters: Sale Date: After 8/23/17; CoStar Rating: 2 stars or less; Property type: Multifamily; Construction status: Existing; Rent type: Market; Year built: max 1999; Class: B, C, and unclassified
# ALL SALES --- Location: Richmond city, Henrico County, Chesterfield County, Hanover County
# Filters: Sale Date: After 8/23/17

noah_sales <- read_csv("data/noah_sales_82322.csv") |> 
  clean_names() |> 
  mutate(category = "NOAH", .before = 1)

all_sales <- read_csv("data/all_sales_82322.csv") |> 
  clean_names() |> 
  mutate(category = "All", .before = 1)

mf_sales <- rbind(noah_sales, all_sales) |> 
  filter(!period %in% c("Search", "2017 Q2", "2017 Q1")) |> 
  mutate(period = case_when(
    period == "2022 Q3 QTD" ~ "2022 Q3",
    TRUE ~ period
  )) |> 
  mutate(across(.cols = 3:26, ~ parse_number(.x))) |> 
  select(category, period, sales = number_of_transactions, volume = sales_volume,
         avg_price = average_sale_price, avg_price_sf = average_price_per_bldg_sf,
         med_price_sf = median_price_per_bldg_sf, sold_units,
         avg_price_unit = average_price_per_unit, med_price_unit = median_price_per_unit) |> 
  mutate(period = as.yearqtr(period, format = "%Y Q%q"))

write_rds(mf_sales, "data/pha_mf_sales.rds")

```

The sale of NOAH properties and ensuing new ownership often leads to rent increases and/or rehabilitation. In some cases, this may take NOAH properties out of market affordability. Over the last five years, NOAH properties have made up well over half of all multifamily property transactions in the region.

```{r}
#| label: fig-noah-sales
#| fig.cap: "Number of multifamily properties sold"

mf_sales <- read_rds("data/pha_mf_sales.rds")

mf_sales_plot <- ggplot(mf_sales,
       aes(x = period,
           y = sales,
           fill = category,
           data_id = sales,
           tooltip = sales)) +
  geom_col(position = "dodge") +
  geom_col_interactive(position = "dodge", size = 2) +
  theme_pha() +
  scale_fill_pha() +
  scale_x_yearqtr(format = "%Y Q%q",
                  breaks = seq(from = min(mf_sales$period), to = max(mf_sales$period), by = 0.25)) +
  labs(title = "Number of multifamily properties sold",
       subtitle = "2017 Q3 to 2022 Q3 (to-date)",
       caption = "**Source:** CoStar Group, Inc.") +
  theme(legend.position = "top",
        axis.text.x = element_text(angle = 90, vjust = 0.5),
        plot.subtitle = element_markdown(margin = margin(t = 5, b = 0, unit = "pt"))) +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = mf_sales_plot,
         height_svg = 4)
  
} else {
  
  mf_sales_plot
  
  }


```

However, NOAH transactions represent a smaller share of total units sold---about one-third since 2017 Q3.

```{r}
#| label: fig-noah-sales-units
#| fig.cap: "Number of multifamily units sold"

mf_unit_plot <- ggplot(mf_sales,
       aes(x = period,
           y = sold_units,
           fill = category,
           data_id = sales,
           tooltip = sales)) +
  geom_col(position = "dodge") +
  geom_col_interactive(position = "dodge", size = 2) +
  theme_pha() +
  scale_fill_pha() +
  scale_x_yearqtr(format = "%Y Q%q",
                  breaks = seq(from = min(mf_sales$period), to = max(mf_sales$period), by = 0.25)) +
  scale_y_continuous(labels = label_comma()) +
  labs(title = "Number of multifamily units sold",
       subtitle = "2017 Q3 to 2022 Q3 (to-date)",
       caption = "**Source:** CoStar Group, Inc.") +
  theme(legend.position = "top",
        axis.text.x = element_text(angle = 90, vjust = 0.5),
        plot.subtitle = element_markdown(margin = margin(t = 5, b = 0, unit = "pt"))) +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = mf_unit_plot,
         height_svg = 4)
  
} else {
  
  mf_unit_plot
  
  }

```

Sales volume for NOAH properties has, on average, been a small fraction of total volume in the region. However, beginning in 2021 Q4, NOAH sales volume rose above \$100 million for the first time since 2018 Q3, which itself was an an outlier. NOAH sales volume hit a new record in 2022 Q1 (over \$178 million), stayed at that level the next quarter, and continues to be well above average to-date in 2022 Q3.

```{r}
#| label: fig-noah-sales-vol
#| fig.cap: "Volume of multifamily sales"

mf_sales_plot <- ggplot(mf_sales,
       aes(x = period,
           y = volume,
           color = category)) +
  geom_line(size = 1) +
  geom_point_interactive(
    aes(data_id = volume,
        tooltip = label_dollar()(volume)),
    size = 0.5) + 
  theme_pha() +
  scale_color_pha() + 
    scale_y_continuous(labels = label_dollar(scale = 0.000001, suffix = "M")) +
  scale_x_yearqtr(format = "%Y Q%q",
                  breaks = seq(from = min(mf_sales$period), to = max(mf_sales$period), by = 0.25)) +
  labs(title = "Volume of multifamily sales",
       subtitle = "2017 Q3 to 2022 Q3 (to-date)",
       caption = "**Source:** CoStar Group, Inc.") +
  theme(legend.position = "top",
        axis.text.x = element_text(angle = 90, vjust = 0.5)) +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = mf_sales_plot,
         height_svg = 4)
  
} else {
  
  mf_sales_plot
  
  }



```

During this timeframe, NOAH properties had an average price per unit below that of all multifamily sales until 2020 Q4. At the end of 2020, the average price per unit of a NOAH property hit a high of \$221,534 --- over \$44,000 more when compared to all multifamily sales. Although both types of sales took a dip following the end of 2020, the higher NOAH price remained until 2021 Q2, when NOAH average price per unit once again went below all sales.

```{r}
#| label: fig-noah-price-unit
#| fig.cap: "Average price per unit"

unit_price <- ggplot(mf_sales,
       aes(x = period,
           y = avg_price_unit,
           color = category))  +
  geom_line(size = 1) +
  geom_point_interactive(
    aes(data_id = avg_price_unit,
        tooltip = label_dollar()(avg_price_unit)),
    size = 0.5) + 
  theme_pha() +
  scale_color_pha() + 
  scale_x_yearqtr(format = "%Y Q%q",
                  breaks = seq(from = min(mf_sales$period), to = max(mf_sales$period), by = 0.25)) +
  scale_y_continuous(labels = label_dollar(scale = 0.001, suffix = "K")) +
  labs(title = "Average price per unit",
       subtitle = "2017 Q3 to 2022 Q3 (to-date)",
       caption = "**Source:** CoStar Group, Inc.") +
  theme(legend.position = "top",
        axis.text.x = element_text(angle = 90, vjust = 0.5),
        plot.subtitle = element_markdown(margin = margin(t = 5, b = 0, unit = "pt"))) +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = unit_price,
         height_svg = 4)
  
} else {
  
  unit_price
  
  }

```

## Manufactured home communities

Manufactured home communities (MHCs) are also a valuable source of NOAH across the region, but are not reliably monitors as traditional multifamily rental housing. As a result, accurate data on supply and rents are more difficult to obtain.

### Supply

In 2016, the Manufactured Home Community Coalition of Virginia (MHCCV) conducted an assessment of all manufactured home communities across the Richmond, Virginia Metropolitan Statistical Area (MSA). That report found 4,735 homes across 54 MHCs in the greater region. Within the primary PHA area^[Chesterfield, Hanover, Henrico, and Richmond.] there are 24 different MHCs, which in all contained at least 2,742 individual manufactured homes. 

::: callout-note

This data includes homes that may be rented, as well as owned. Regardless, residents in MHCs rent the lot on which their home resides. This leaves many manufactured home community residents who own their homes in a precarious position should a community owner decide to sell or redevelop the property. Manufactured homes are not easily moved once installed, leaving many families forced to abandon their homes and seek new and more expensive housing elsewhere.

:::

```{r}
#| label: fig-mhc-supply
#| fig.cap: "Units in manufactured home communities by locality"

mhccv <- read_csv("data/mhccv_data.csv") |> 
  clean_names()

names <- c("Chesterfield", "Hanover", "Henrico", "Richmond")

mhc_cnt <- mhccv |> 
  filter(county %in% names) |> 
  group_by(county) |> 
  summarise(units = sum(units),
            mhcs = n())

mhc_plot <- ggplot(mhc_cnt,
       aes(y = reorder(county, units),
           x = units, 
           fill = county,
           data_id = units,
           tooltip = label_comma()(units))) +
  geom_col() + 
  geom_col_interactive(size = 2) +
  theme_pha() +
  scale_fill_pha() +
  scale_x_continuous(labels = label_comma()) +
  labs(title = "Units in manufactured home communities by locality",
       caption = "**Source:** Manufactured Home Community Coalition of Virginia.") +
  theme(plot.title = element_markdown(margin = margin(b = 20, unit = "pt"))) +
  flip_gridlines() +
  add_zero_line(axis = "x")

if (knitr::is_html_output()) {
  
  girafe(ggobj = mhc_plot,
         height_svg = 4)
  
} else {
  
  mhc_plot
  
  }



```
Chesterfield County has the largest supply of homes in MHCs (1,543), which is about half the total number of subsidized rentals also in the county. Hanover and Richmond both have near 500 units in MHCs, while Henrico only has one MHC with 230 units.

### Locations

The majority of manufactured home community units are located along the Route 1 corridor. In areas where commercial and mixed-use development has accelerated, many of these properties are well-positioned for a change to a "higher and better" use. This redevelopment potential, while often in line with broader planning goals, is a threat to the long-term stability of MHCs.

```{r}
#| label: mhc-map
#| eval: false

mhc <- st_read("https://services1.arcgis.com/Hp6G80Pky0om7QvQ/arcgis/rest/services/Mobile_Home_Parks/FeatureServer/0/query?outFields=*&where=1%3D1&f=geojson")

pha_mhc <- mhc |>
  subset(COUNTYFIPS %in% pha)

write_rds(pha_mhc, "data/pha_mhc.rds")

```

```{r}
#| label: fig-mhc-map
#| fig.cap: "Manufactured housing communities in Richmond region"

pha_mhc <- read_rds("data/pha_mhc.rds") |>
  filter(STATUS != "CLOSED") |> 
  mutate(SIZE = str_to_sentence(SIZE))

if (knitr::is_html_output()) {
  
  mapview(
    pha_mhc,
    zcol = "SIZE",
    legend = T,
    popup = F,
    burst = F,
    layer.name = "Units in park"
  )
  
  } else {

  ggplot() +
    rr_basemap +
    geom_sf(pha_mhc, mapping = aes(color = SIZE), show.legend = "point") +
    scale_color_pha() +
    scale_x_continuous(expand = expansion(mult = 0),
                       limits = c(-8660000, -8585000)) +
    scale_y_continuous(expand = expansion(mult = 0),
                       limits = c(4466959, 4548000)) +
    labs(title = "Manufactured housing communities in Richmond region",
         color = "Units in park",
         caption = "**Source:** Manufactured Home Community Coalition of Virginia.") +
    theme_pha() +
    theme(plot.title = element_markdown(margin = margin(b = 10, unit = "pt")),
          panel.grid = element_blank(),
          axis.title = element_blank(),
          axis.text = element_blank(),
          legend.position = "right",
          legend.title = element_text(),
          legend.key = element_blank())
    
  }

```

::: callout-tip

### Notable changes since 2020

In September 2020, project:HOMES, [acquired](https://www.projecthomes.org/bermuda-estates) a 52 unit manufactured home community called Bermuda Estates in Chesterfield County. Since acquiring the community, project:HOMES has made significant infrastructure improvements, replaced some units in disrepair, and constructed a community center. The nonprofit plans to continue investments and preserve the park as a high-quality, low-cost neighborhood.

Suburban Village, Chesterfield County's third largest MHC with almost 250 units, was [purchased](https://richmond.com/news/local/maryland-based-investment-firm-buys-chesterfield-mobile-home-park-for-22-5-million/article_41bd2656-0466-5643-9285-2715e88fee40.html) by Maryland-based Horizon Land Management in August 2021 for \$22.5 million. The park was previously under the same local ownership since 1986. More than 35 potential buyers expressed interest.

Shady Hill Mobile Home Park, home to more than 100 families, was purchased by a Charlottesville-based development firm for \$5.1 million in August 2022. While complete redevelopment is likely, exact plans and timing are not known.

:::
