==> part-3-2.qmd <==
# Impact of housing costs on household budgets {#part-3-2}

```{r}
#| label: setup

library(tidyverse)
library(httr)
library(glue)
library(readxl)
library(janitor)
library(stats)
library(lubridate)
library(scales)
library(ggiraph)
library(ggtext)
library(hdatools)

rr <- c("51760", "51085", "51087", "51041",
        "51145", "51075", "51127", "51036")
```

```{r}
#| label: chas-pull
#| eval: false

years <- 2015:2019

sumlev <- "050"

dir.create(glue("data/{sumlev}"))

walk(years, ~{

  url <- glue("https://www.huduser.gov/PORTAL/datasets/cp/{.x - 4}thru{.x}-{sumlev}-csv.zip")

  file <- basename(url)

  path <- file.path("data", sumlev, file)

  if (!file.exists(path)) {
    GET(url, write_disk(path, overwrite = TRUE), progress(type = "down"))
  }

  print(glue("Unzipping {.x}..."))
  unzip(path, exdir = file.path("data", sumlev, .x))

})

# Tables to get
tables <- c(7, 9, paste0(18, LETTERS[1:3]))

# Go through and write out the various tables
walk(tables, function(table) {

  mytable <- purrr::map_df(years, function(year) {

    # Identify the year-specific folder
    path <- file.path("data", "050", year)

    # Find the file - it may be buried so use recursive = TRUE
    file <- list.files(path, pattern = glue("Table{table}.csv"), recursive = TRUE)

    # Read in the file quietly
    raw <- read_csv(file.path(path, file), col_types = cols())

    # Clean it up
    cleaned <- raw |>
      clean_names() |>
      mutate(fips = substr(geoid, 8, 12)) |>
      separate(name, into = c("county", "state"), sep = ",") |>
      filter(st == "51") |>
      pivot_longer(starts_with("T"),
                   names_to = "code",
                   values_to = "value") |>
      mutate(id = str_extract(code, "\\d+$"),
             type = str_extract(code, "est|moe")) |>
      select(-code) |>
      pivot_wider(names_from = type, values_from = value) |>
      rename(Estimate = est, MOE = moe) |>
      mutate(Code := glue("T{table}_est{id}"),
             Year = year) |>
      select(Code, Year, Estimate, MOE, everything(), -id) |>
      mutate(fips = case_when(
        fips == "51515" ~ "51019",
        TRUE ~ fips
      )) |>
      mutate(county = case_when(
        county == "Bedford city" ~ "Bedford County",
        TRUE ~ county
      )) |>
      subset(fips %in% rr)

    # Account for different data dictionaries
    # Find the data dictionary in the appropriate folder
    dict_path <- list.files(path, pattern = "dictionary", recursive = TRUE, full.names = TRUE)

    # Read in the data dictionary and merge
    dict <- read_excel(dict_path,
                       sheet = glue("Table {table}"))

    cleaned_with_dict <- cleaned %>%
      left_join(dict, by = c("Code" = "Column Name"))

    cleaned_with_dict

  })

  file_name <- glue("Table{table}_2015to2019.csv")

  message(glue("Writing file {file_name}..."))

  write_csv(mytable, glue("data/{file_name}"))

})

# Consider adding script to remove the "data/050" file folder of zips and tables to create space.

# file.remove()


```

```{r}
#| label: chas-data
#| eval: false

cb_7 <- read_csv("data/Table7_2015to2019.csv") |>
  clean_names() |>
  filter(line_type == "Detail") |>
  select(year, estimate, moe, county, fips, tenure, household_income, household_type, cost_burden) |>
  mutate(tenure = case_when(
    tenure == "Owner occupied" ~ "Homeowner",
    tenure == "Renter occupied" ~ "Renter"
  )) |>
  mutate(household_income = case_when(
    household_income == "household income is less than or equal to 30% of HAMFI" ~ "30% AMI or less",
    household_income == "household income is greater than 30% but less than or equal to 50% of HAMFI" ~ "31 to 50% AMI",
    household_income == "household income is greater than 50% but less than or equal to 80% of HAMFI" ~ "51 to 80% AMI",
    household_income == "household income is greater than 80% but less than or equal to 100% of HAMFI" ~ "81 to 100% AMI",
    household_income == "household income is greater than 100% of HAMFI" ~ "101% AMI or greater"
  )) |>
  mutate(household_type = case_when(
    household_type == "household type is elderly family (2 persons, with either or both age 62 or over)" ~ "Elderly family",
    household_type == "household type is small family (2 persons, neither person 62 years or over, or 3 or 4 persons)" ~ "Small family",
    household_type == "household type is large family (5 or more persons)" ~ "Large family",
    household_type == "household type is elderly non-family" ~ "Elderly non-family",
    household_type == "other household type (non-elderly non-family)" ~ "Non-elderly non-family"
  ))  |>
  mutate(cost_burden = case_when(
    cost_burden == "housing cost burden is less than or equal to 30%" ~ "Not cost-burdened",
    cost_burden == "housing cost burden is greater than 30% but less than or equal to 50%" ~ "Cost-burdened",
    cost_burden == "housing cost burden is greater than 50%" ~ "Severely cost-burdened",
    cost_burden == "housing cost burden not computed (household has no/negative income)" ~ "No or negative income"
  )) |>
  mutate(cb_group = case_when(
    cost_burden == "Cost-burdened" ~ "Cost-burdened",
    cost_burden == "Severely cost-burdened" ~ "Cost-burdened",
    TRUE ~ cost_burden
  ))

write_rds(cb_7, "data/cb_7.rds")

```

## Cost burden

When incomes don't rise along with housing costs, we can expect an increase in the number of cost-burdened households who pay more than 30 percent of their gross income on basic housing expenses. Since 2015, cost burden levels in the region decreased for some groups, while increased for others.

Data in this section come from the Comprehensive Housing Affordability Strategy (CHAS) dataset published by the U.S. Department of Housing and Urban Development. CHAS estimates are a custom tabulation of American Community Survey responses. As of October 2022, the most recent CHAS data is for the 2015-2019 5-year period.

Unless otherwise noted, all plots on this page combine data from Chesterfield County, Hanover County, Henrico County, and Richmond city.

### Cost burden by tenure

The number of cost-burdened homeowners across the region has declined significantly since 2015, particularly in Chesterfield and Henrico counties. Hanover County and Richmond city saw smaller decreases, but the total "loss" of cost-burdened homeowners in the region still exceeded 7,200.

Meanwhile, the total number of cost-burdened renter households increased by almost 1,900, with only Hanover County seeing a small decline. Much of this growth in renter cost burden was focused in Chesterfield County and Richmond city.

```{r}
#| label: fig-cb-locality
#| fig.cap: "Cumulative change in cost-burdened households by tenure"

cb_7 <- read_rds("data/cb_7.rds")

pha_localities <- c("Richmond city", "Chesterfield County", "Hanover County", "Henrico County")

cb_locality <- cb_7 |> 
  filter(county %in% pha_localities,
         cb_group == "Cost-burdened") |> 
  group_by(year, county, tenure) |> 
  summarise(estimate = sum(estimate)) |> 
  group_by(county, tenure) |> 
  mutate(change = estimate - lag(estimate)) |> 
  drop_na() |> 
  mutate(run_diff = cumsum(change))
  
cb_diff <- ggplot(cb_locality,
       aes(x = year, 
           y = run_diff, 
           fill = tenure,
           data_id = run_diff,
           tooltip = run_diff)) +
  geom_col(position = "dodge") +
  geom_col_interactive(position = "dodge", size = 2) +
  theme_pha() +
  scale_fill_pha() +
  facet_wrap(vars(county)) +
  scale_y_continuous(labels = label_comma()) +
  labs(title = "Cumulative change in cost-burdened households by tenure",
       subtitle = "2015 to 2019",
       fill = "Tenure",
       caption = "**Source:** U.S Department of Housing and Urban Development,<br>Comprehensive Housing Affordability Strategy (CHAS), Table 7.") +
  theme(legend.position = "right") +
  add_zero_line(axis = "y")


if (knitr::is_html_output()) {
  
  girafe(ggobj = cb_diff,
         height_svg = 4) 
  
} else {
  
  cb_diff
  
  }

```

### Cost burden by income

Homeowners above 80 percent AMI saw the largest declines in cost burden since 2015. This is likely due to rising incomes among homeowners with relatively fixed housing costs. Although renters with higher incomes were seeing growing cost burden from 2015 to 2018, 2019 estimates show an abrupt shift in trends. Renters making less than 50 percent AMI saw an increase in cost burden from 2015 of nearly 4,000 households. The number of households making greater than 50 percent AMI saw an overall decrease in cost burden from 2015 estimates --- reversing the increases seen in 2018.

```{r}
#| label: fig-cb-income
#| fig.cap: "Cumulative change in cost-burdened households by income and tenure"

cb_income <- cb_7 |> 
  filter(county %in% pha_localities,
         cb_group == "Cost-burdened") |> 
  group_by(year, household_income, tenure) |> 
  summarise(estimate = sum(estimate)) |> 
  group_by(household_income, tenure) |> 
  mutate(change = estimate - lag(estimate)) |> 
  drop_na() |> 
  mutate(run_diff = cumsum(change)) |> 
  ungroup() |> 
  mutate(household_income = fct_relevel(household_income,
                                        "30% AMI or less",
                                        "31 to 50% AMI",
                                        "51 to 80% AMI",
                                        "81 to 100% AMI",
                                        "101% AMI or greater"))

cb_inc_plot <- ggplot(cb_income,
       aes(x = year, 
           y = run_diff, 
           fill = tenure,
           data_id = run_diff,
           tooltip = label_comma()(run_diff))) +
  geom_col(position = "dodge") +
  geom_col_interactive(position = "dodge", size = 2) +
  theme_pha() +
  scale_fill_pha() +
  facet_wrap(~household_income, 
             nrow = 1, 
             labeller = label_wrap_gen(width = 12, multi_line = TRUE)) +
  scale_y_continuous(labels = label_comma()) +
  labs(title = "Cumulative change in cost-burdened households<br>by income and tenure",
       subtitle = "2015 to 2019",
       fill = "Tenure",
       caption = "**Source:** U.S Department of Housing and Urban Development,<br>Comprehensive Housing Affordability Strategy (CHAS), Table 7.") +
  theme(legend.position = "right",
        axis.text.x = element_text(angle = 90, hjust = 0, vjust = 0)) +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = cb_inc_plot,
         height_svg = 4) 
  
} else {
  
  cb_inc_plot
  
  }
```

However, the significant and unexpected drop among cost-burdened renters below 30 percent AMI from 2017 to 2018 deserves further explanation. The increase in 2019 appears to be more aligned with our sense of the rental housing market over the past few years.   

The plot below shows the ACS estimates of renter households by cost burden from 2016 to 2020. There is a steady decline in the number of cost-burdened low-income renters (under \$35,000); however, this corresponds to an increasing number of cost-burdened renters with incomes between \$35,000 and \$75,000.

```{r}
#| label: cb-income-acs-data
#| eval: false

years_acs <- 2015:2020

pha_geoid <- c("51085", "51760", "51087", "51041")

b25074_raw <- map_dfr(years_acs, function(yr){
  b25074_pull <- get_acs(
    geography = "county",
    table = "B25074",
    state = "VA",
    year = yr
  ) |>
    mutate(year = yr)
})

b25074_vars <- load_variables(2020, dataset = "acs5") |>
  filter(str_detect(name, "B25074"))

b25074_vars_cleaned <- b25074_vars |>
  separate(label, into = c("est", "tot", "inc", "cb"), sep = "!!") |>
  select(variable = name, inc, cb) |>
  mutate(across(.fns = ~str_remove_all(.x, ":"))) |>
  drop_na() |>
  mutate(cb_group = case_when(
    str_detect(cb, "30.0|35.0|40.0|50.0") ~ "Cost-burdened",
    str_detect(cb, "20.0|25.0") ~ "Not cost-burdened",
    TRUE ~ "Not computed"
  ))

b25074_data <- b25074_raw |>
  filter(GEOID %in% pha_geoid) |>
  right_join(b25074_vars_cleaned, by = "variable") |>
  filter(!year == 2015) |>
  group_by(GEOID, year, inc, cb_group) |>
  summarise(estimate = sum(estimate)) |>
  group_by(GEOID, inc, cb_group) |>
  mutate(change = estimate - lag(estimate)) |>
  ungroup() |>
  mutate(inc = fct_relevel(inc,
                           "Less than $10,000",
                           "$10,000 to $19,999",
                           "$20,000 to $34,999",
                           "$35,000 to $49,999",
                           "$50,000 to $74,999",
                           "$75,000 to $99,999",
                           "$100,000 or more"),
         cb_group = fct_relevel(cb_group,
                                "Cost-burdened",
                                "Not cost-burdened",
                                "Not computed"))

write_rds(b25074_data, "data/b25074_data.rds")

```

```{r}
#| label: fig-cb-income-acs-est
#| fig.cap: "Renter households by income and cost burden"

b25074_data <- read_rds("data/b25074_data.rds") |> 
  group_by(year, inc, cb_group) |> 
  summarise(estimate = sum(estimate))

renter_cb <- ggplot(b25074_data,
       aes(x = year, 
           y = estimate, 
           fill = cb_group,
           data_id = estimate,
           tooltip = label_comma()(estimate))) +
  geom_col(position = "dodge") +
  geom_col_interactive(position = "dodge", size = 2) +
  theme_pha() +
  scale_fill_pha() +
  facet_wrap(~inc, nrow = 1, labeller = label_wrap_gen(width = 12, multi_line = TRUE)) +
  labs(title = "Renter households by income and cost burden",
       subtitle = "2016 to 2020",
       caption = "**Source:** U.S. Census Bureau, American Community Survey, Table B25074.") +
  scale_y_continuous(labels = label_comma()) +
  theme(axis.text.x = element_text(size = 8, angle = 90, vjust = 0.5),
        legend.position = "bottom") +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = renter_cb,
         height_svg = 4) 
  
} else {
  
  renter_cb
  
  }

```

Nearly all cost-burdened renter households have incomes below \$75,000. Filtering for just those estimates, the plot below shows the net annual change from 2016 to 2020. The significant decrease from 2019 to 2020 (1,057) is well beyond the range from previous changes, and may also be due in part to lower ACS response rates among lower-income households during the COVID-19 pandemic.

```{r}
#| label: fig-cb-income-acs-yoy
#| fig.cap: "Year-over-year change in cost-burdened renter households"

b25074_yoy <- b25074_data |>
  group_by(year, inc, cb_group) |> 
  filter(cb_group == "Cost-burdened",
         !inc %in% c("$75,000 to $99,999", "$100,000 or more")) |> 
  summarise(sum = sum(estimate)) |> 
  group_by(year) |> 
  summarise(sum = sum(sum)) |> 
  mutate(yoy = sum - lag(sum)) |> 
  drop_na()

acs_yoy_plot <- ggplot(b25074_yoy,
       aes(x = year, 
           y = yoy,
           data_id = yoy,
           tooltip = label_comma()(yoy))) +
  geom_col(fill = "#5bab8e") +
  geom_col_interactive(fill = "#5bab8e", size = 2) +
  theme_pha() + 
  scale_fill_gradientn(colours = c("#CAE3C2", "#74BA91", "#519B8D", "#2B4258")) +
  labs(title = "Year-over-year change in cost-burdened renter households",
       subtitle = "Household income below $75,000 only",
       caption = "**Source:** U.S. Census Bureau, American Community Survey, Table B25074.") +
  scale_y_continuous(labels = label_number(big.mark = ",", style_positive = "plus", style_negative = "minus")) +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = acs_yoy_plot,
         height_svg = 4)
  
} else {
  
  acs_yoy_plot +
    geom_text(aes(y = yoy - sign(yoy)/0.02,
                  label = label_comma()(yoy)),
              color = "white")
  
  }

```

In summary, since the *total* number of renter households in the region has not changed significantly from 2016 to 2020, and because the supply of deeply affordable rental housing has not increased, the estimated decline in low-income cost-burdened renters is likely due to a combination of increasing average incomes "re-sorting" households into higher income categories, as well as pandemic data collection challenges.

### Cost burden by household type

Small families and non-elderly, non-family homeowner households saw the largest decreases in cost burden across all four localities. Among renters, only small family households are now less likely to be cost-burdened, but this change (-765) is an order of magnitude smaller than the decrease for homeowner small families (-6,525).

Net increases in cost-burden were almost entirely contained to elderly non-family and elderly family households. There are now more than 4,600 additional cost-burdened households in these groups, including both homeowners and renters.

```{r}
#| label: fig-cb-hh
#| fig.cap: "Change in cost-burdened households by household type"

cb_hh_pha <- cb_7 |> 
  subset(county %in% pha_localities) |> 
  group_by(year, tenure, household_type, cb_group) |> 
  summarise(estimate = sum(estimate)) |> 
  pivot_wider(
    names_from = year,
    values_from = estimate
  ) |> 
  clean_names() |> 
  transform(difference = (x2019 - x2015)) |> 
  filter(cb_group == "Cost-burdened") |> 
  mutate(household_type = fct_reorder(household_type, difference, .desc = TRUE))

hh_plot <- ggplot(cb_hh_pha,
       aes(x = household_type,
           y = difference,
           fill = household_type,
           data_id = difference,
           tooltip = label_comma()(difference))) +
  geom_col() +
  geom_col_interactive(size = 2) +
  facet_wrap(~tenure) +
  theme_pha() +
  scale_fill_pha() +
  labs(title = "Change in cost-burdened households by household type",
       subtitle = "2015 to 2019",
       caption = "**Source:** U.S Department of Housing and Urban Development,<br>Comprehensive Housing Affordability Strategy (CHAS), Table 7.") +
  scale_y_continuous(labels = label_number(big.mark = ",", style_positive = "plus", style_negative = "minus")) +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 10)) +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = hh_plot,
         height_svg = 4)
  
} else {
  
  hh_plot
  
  }

```

## Mortgage delinquency and foreclosure

Since the Great Recession, mortgage delinquency of 90 days or more has been on a steady decline across the region ---reaching the decade's lowest rates throughout much of 2020 and 2021. Pandemic mortgage relief measures laid out in the CARES Act led to a significant forbearance program, wherein homeowners with federally-backed mortgages could enter into forbearance for a year. The decrease in delinquency can be greatly attributed to these measures which stipulated that loans in forbearance would not be reported as delinquent.

According to some researchers, this program also led to loans in delinquency prior to the pandemic entering into forbearance as well.^[Haughwout, Lee, Scally, and van der Klaauw, 2020. https://libertystreeteconomics.newyorkfed.org/2020/11/following-borrowers-through-forbearance/] Interestingly, Hanover County saw a spike in mortgage delinquency during 2018, but has since declined to the lowest rate (0.2 percent) among all localities as of December 2021.

```{r}
#| label: fig-mort-del
#| fig.cap: "Mortgage delinquency rate by locality"

mort_del <- read_csv("data/mortgage_del.csv") |>
  pivot_longer(
    cols = starts_with("2"),
    names_to = "period",
    values_to = "estimate"
  )|>
  mutate(period = ym(period)) |>
  filter(State == "VA") |>
  subset(Name %in% pha_localities) |>
  mutate(estimate = (estimate/100))

mortd_plot <- ggplot(mort_del,
       aes(x = period,
           y = estimate,
           color = Name)) +
  geom_line(size = 1) +
  geom_point_interactive(
    aes(data_id = estimate,
        tooltip = label_percent()(estimate)),
    size = 0.5) + 
  theme_pha() +
  scale_color_pha() + 
  scale_y_continuous(labels = label_percent(accuracy = 0.1), expand = c(0,0), limits = c(0,0.055)) +
  scale_x_date(date_breaks = "2 years", date_labels = "%Y") +
  labs(title = "Mortgage delinquency rate by locality",
       subtitle = "Mortgages 90 or more days delinquent: 2008 to 2021",
       color = "Locality",
       caption = "**Source:** Consumer Financial Protection Bureau, Mortgage Performance Trends.") +
  theme(legend.position = "right") +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = mortd_plot,
         height_svg = 4)
  
} else {
  
  mortd_plot
  
}


```

With the moratorium on residential foreclosures having come to an end on June 30, 2022, the region may see increasing mortgage delinquency rates in the coming years.

## Eviction filings and judgements

Richmond's elevation to national prominence due to its eviction rate spurred state-level responses to address the eviction crisis across the Commonwealth. From 2017 to 2019, the region saw small declines in the number of eviction filings. The City of Richmond saw a 14 percent decrease in average annual filings, while eviction judgements only decreased by 8 percent.

::: callout-tip

### Defining evictions

For this section, we define eviction *filings* as the number of lawsuits generated by landlords against tenants to begin eviction proceedings. Eviction *judgements* are the subsequent court orders for tenants to vacate their apartment. Not every eviction case results in a judgement, and not every judgement results in a formal eviction carried out by local sheriff's deputies.

:::

The eviction landscape changed dramatically during the COVID-19 pandemic when the Centers for Disease Control imposed a nationwide federal moratorium on residential evictions in September 2020. In Virginia, Governor Northam requested from the state's Supreme Court a stay on evictions preceding the nationwide moratorium several times.

```{r}
#| label: fig-evictions
#| fig.cap: "Evictions filings and judgements by locality"

evictions <- read_csv("data/evictions_17_22.csv") %>% 
  pivot_longer(3:142,
               names_to = "data",
               values_to = "value") |>
  mutate(type = case_when(
    str_detect(data, "Filings") ~ "Filings",
    str_detect(data,"Evictions") ~ "Eviction judgements"
  )) |>
  mutate(date = str_sub(data, -7, -1)) |>
  mutate(date = ym(date)) |>
  mutate(year = year(date)) |>
  mutate(type = fct_relevel(type, "Filings", "Eviction judgements"))

eviction_colors <- setNames(c("#a6cccc", "#be451c"),
                            nm = c("Filings", "Eviction judgements"))

evict_plot <- ggplot(evictions,
       aes(x = date,
           y = value,
           fill = type,
           data_id = value,
           tooltip = label_comma()(value))) +
  geom_col(position = "dodge") +
  geom_col_interactive(position = "dodge", size = 2) +
  facet_wrap(~Jurisdiction) +
  theme_pha() +
  scale_x_date(labels = label_date_short(), breaks = breaks_width("6 months")) +
  scale_y_continuous(labels = label_comma(), expand = expansion(mult = 0)) +
  labs(title = "Evictions filings and judgements",
       subtitle = "January 2017 through October 2022",
       caption = "**Source:** RVA Eviction Lab, Through October 2022") +
  scale_fill_manual(values = eviction_colors) +
  theme(legend.position = "bottom",
        axis.text.x = element_text(size = 7)) +
  add_zero_line(axis = "y")


if (knitr::is_html_output()) {
  
  girafe(ggobj = evict_plot,
         height_svg = 4) 
  
} else {
  
  evict_plot
  
}

```

These measures led to dramatic decreases in both the number of filings and eviction judgements across the region. However, the eviction moratorium's official end in Virginia on June 30, 2022, brings about concerns among advocates and service providers over a potential wave of evictions and homelessness in the coming months. The last quarter of 2022 has already seen major increases in evictions that are well on their way to exceeding pre-pandemic levels.

```{r}
#| label: fig-evict-avg
#| fig.cap: "Average annual eviction filings and orders by locality"

eviction_avg <- evictions |>
  group_by(Jurisdiction, year, type) |>
  summarise(avg = mean(value)) |>
  mutate(type = fct_relevel(type, "Filings", "Eviction judgements"))

evict_avg_plot <- ggplot(eviction_avg,
       aes(x = year,
           y = avg,
           fill = type,
           data_id = avg,
           tooltip = label_comma(accuracy = 1)(avg))) +
  geom_col(position = "dodge") +
  geom_col_interactive(position = "dodge", size = 2) +
  facet_wrap(~Jurisdiction) +
  theme_pha() +
  labs(title = "Average annual eviction filings and orders by locality",
       subtitle = "2017 to 2022 YTD",
       caption = "**Source:** RVA Eviction Lab, Through October 2022.") +
  scale_y_continuous(labels = label_comma()) +
  scale_fill_manual(values = eviction_colors) +
  scale_x_discrete(limits = 2017 + 0:5) +
  theme(legend.position = "bottom") +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = evict_avg_plot,
         height_svg = 4)
  
} else {
  
  evict_avg_plot
  
}

```

Eviction filings should continue to be monitored over the coming months. The [RVA Eviction Lab](https://rampages.us/rvaevictionlab/) has been at the forefront of this data collection and analysis, and will continue to be a resource for the region in understanding the increasing risks for renters with renter protections and resources coming to an end.

## Housing Resource Line

On September 1, 2020, PHA launched the [Housing Resource Line](https://pharva.com/housing-hotline/) to help residents across Central Virginia in need of housing. As of November 2022, the hotline has fielded nearly 17,000 calls from people across the region---from rural Goochland County to the City of Richmond.

Call volume has remained steady over since the line's launch. Call volume has not dropped below 500 calls per month since March 2021.

```{r}
#| label: fig-call-vol
#| fig.cap: "Housing Resource Line monthly call volume"

call_vol <- read_csv("data/hrl_vol.csv") |>
  mutate(date = ym(date))

# sum(call_vol$volume)
#
# 14,838

call_plot <- ggplot(call_vol,
       aes(x = as.Date(date),
           y = volume,
           fill = volume,
           data_id = volume,
           tooltip = volume))  +
  geom_col() +
  geom_col_interactive(size = 2) +
  scale_fill_gradientn(colours = c("#CAE3C2", "#74BA91", "#519B8D", "#2B4258")) +
  theme_pha() +
  scale_x_date(labels = label_date_short(), breaks = breaks_width("3 months")) +
  labs(title = "Housing Resource Line monthly call volume",
       subtitle = "September 2020 through October 2022",
       caption = "**Source:** Partnership for Housing Affordability.") +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = call_plot,
         height_svg = 4) 
  
} else {
  
  call_plot
  
}

```

The majority of calls were for rental options (36 percent) and financial assistance (21 percent). The two other largest share of calls were for an option not listed (17 percent) and homelessness (12 percent).

```{r}
#| label: fig-call-topic
#| fig.cap: "Housing Resource Line volume by call topic"

call_top <- read_csv("data/hrl_need.csv") |>
  mutate(need = case_when(
    need == "Other - specify in \"Issues\" tab" ~ "Other",
    TRUE ~ need
  )) |>
  transform(percent = ave(vol,FUN = prop.table)) |>
  filter(percent > 0.01)

top_plot <- ggplot(call_top,
       aes(y = reorder(need, vol),
           x = percent,
           data_id = percent,
           tooltip = label_percent(accuracy = 0.1)(percent))) +
  geom_col(fill = "#5bab8e") +
  geom_col_interactive(fill = "#5bab8e", size = 2) +
  scale_x_continuous(labels = label_percent(),
                     expand = expansion(mult = c(0,0.05))) +
  theme_pha() + 
  
  labs(title = "Housing Resource Line volume by call topic",
       subtitle = "September 2020 through October 2022",
       caption = "Note: Does not include topics below 1 percent of call volume.<br>**Source:** Partnership for Housing Affordability.") +
  theme(axis.title = element_blank(),
        legend.position = "none",
        panel.background = element_blank(),
        panel.grid.major.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey95",
                                          size = 0.05),
        axis.ticks = element_blank()) +
  scale_x_continuous(labels = label_percent()) +
  flip_gridlines() +
  add_zero_line(axis = "x")

if (knitr::is_html_output()) {
  
  girafe(ggobj = top_plot,
         height_svg = 4)
  
} else {
  
  top_plot
  
}

```

Unsurprisingly, there is an increase in homelessness calls during the colder months. PHA staff note that there is an overall increase in calls during the summer months---specifically in regards to people searching for rental options.

This uptick in rental option calls could be directly related to lease non-renewals as landlords sought to increase rents (potentially to recoup losses from the pandemic) and the increasing demand for student rental options ahead of the fall semester.

## Homelessness

### Point-in-Time counts

From 2011 to 2019, the overall count of persons experiencing homelessness across the Greater Richmond Continuum of Care (GRCoC) had been in decline.^[GRCoC covers City of Richmond, and the counties of Charles City, Chesterfield, Goochland, Hanover (including the town of Ashland), Henrico, New Kent, and Powhatan]. But when the COVID-19 pandemic hit, the count jumped---going from 497 in 2019 to 834 in 2021, a 68 percent increase.

The [Urban Institute recently highlighted](https://www.urban.org/sites/default/files/publication/104529/richmond-virginia-response-to-homelessness-during-the-covid-19-pandemic_0.pdf) Homeward's (the region's planning and coordinating organization for the GRCoC) efforts to address homelessness during the pandemic. Their response measures served as best practice examples in preventing high transmission rates among people experiencing homelessness as well as direct service staff.

But the challenges of reducing homelessness during the pandemic were laid bare. With an eviction moratorium, rental vacancy rates reached record lows---leaving many seeking rental options with little to none. In addition, providers have also referenced landlords setting high security deposits.

```{r}
#| label: pit-data
#| eval: false

years_pit <- as.character(2007:2021)

path <- "data/2007-2021-PIT-Counts-by-CoC.xlsx"

full_set <- map_dfr(years_pit, function(yr) {

pit_set <- read_excel(path, sheet = yr) |>
    clean_names()

suffix_to_remove <- paste0("_", yr)

  if ("co_c_category" %in% names(pit_set)) {
    pit_renamed <- pit_set %>%
      rename_with(~str_remove(.x, suffix_to_remove)) |>
      mutate(year = yr) %>%
      select(co_c_number:co_c_name, co_c_category, type_of_count, year, everything()) |>
      mutate(across(-c(co_c_number:year), .fns = as.numeric))
  } else {
    pit_renamed <- pit_set %>%
      rename_with(~str_remove(.x, suffix_to_remove)) |>
      mutate(year = yr, co_c_category = NA_character_,
             type_of_count = NA_character_) |>
      select(co_c_number:co_c_name, co_c_category, type_of_count, year, everything()) |>
      mutate(across(-c(co_c_number:year), .fns = as.numeric))
  }

  pit_renamed

})

full_set_longer <- full_set %>%
  pivot_longer(
    cols = overall_homeless:unsheltered_homeless_parenting_youth_under_25_multiple_races,
    names_to = "category",
    values_to = "value"
  )

rva_coc <- full_set_longer |>
  filter(co_c_number == "VA-500") |>
  filter(str_detect(category, "overall_homeless"))

# Add 2022 PIT count value (directly from Homeward)

rva_pit <- rva_coc |>
  filter(category == "overall_homeless") |>
  add_row(year = "2022", value = 697)

write_rds(rva_pit, "data/rva_pit.rds")

```

```{r}
#| label: fig-pit-plot
#| fig.cap: "Greater Richmond CoC Point-in-Time count"

rva_pit <- read_rds("data/rva_pit.rds")

pit_plot <- ggplot(rva_pit,
       aes(x = year,
           y = value,
           fill = value,
           data_id = value,
           tooltip = label_comma()(value))) +
  geom_col() +
  geom_col_interactive(size = 2) +
  scale_fill_gradientn(colours = c("#CAE3C2", "#74BA91", "#519B8D", "#2B4258")) +
  theme_pha() +
  labs(title = "Greater Richmond Continuum-of-Care Point-in-Time count",
       subtitle = "2007 to 2022",
       caption = "**Sources:** U.S. Department of Housing and Urban Development, CoC Homeless Populations and<br>Subpopulations Reports; Homeward 2022 Winter PIT Count Preliminary Data.") +
  scale_y_continuous(labels = label_comma()) +
  add_zero_line(axis = "y")
  
if (knitr::is_html_output()) {
  
  girafe(ggobj = pit_plot,
         height_svg = 4) 
  
} else {
  
  pit_plot
  
}

```

### Students experiencing homelessness

The McKinney-Vento Education for Homeless Children and Youth (EHCY) Program collects data on students experiencing homelessness, which often can paint a different picture of homelessness when compared to the Point-in-Time counts. In the region, school divisions have been seeing varying numbers, but between the 2018-2019 and 2019-2020 school years students experiencing homelessness have declined across all school divisions.

::: callout-tip

#### Defining student homelessness

Homeless children counted under the McKinney-Vento program are defined as “individuals who lack a fixed, regular, and adequate nighttime residence.” This includes children who are doubled-up with another households or living in motels, along with those living in shelters, vehicles, public areas, and other unsuitable places. This is more expansive than the definition used for PIT counts.

:::

The most notable declines in student homelessness have been seen in the Richmond Public School system, where the number of students experiencing homelessness have declined by 40 percent from 2017-2018 to 2019-2020. Given the pandemic and virtual learning environments, upcoming McKinney-Vento data through the 2021-2022 school year may need require extra context.

```{r}
#| label: fig-mkv
#| fig.cap: "Enrolled students experiencing homelessness by school year"

div <- c(21, 42, 43, 123)

mkv <- read_csv("data/mkv.csv") |>
  subset(Division %in% div)

mkv_plot <- ggplot(mkv,
       aes(x = year,
           y = students,
           fill = Name,
           data_id = students,
           tooltip = label_comma()(students))) +
  geom_col() +
  geom_col_interactive(size = 2) +
  facet_wrap(~Name) + 
  theme_pha() +
  scale_fill_pha() + 
  labs(title = "Enrolled students experiencing homelessness by school year",
       subtitle = "2016-2017 to 2019-2020",
       caption = "**Source:** William & Mary School of Education, Project HOPE - Virginia.") +
  scale_y_continuous(labels = label_comma()) +
  add_zero_line(axis = "y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = mkv_plot,
         height_svg = 4) 
  
} else {
  
  mkv_plot
  
}

```
