==> part-4-5.qmd <==
# Town of Ashland {#part-4-5}

This chapter is a summary of the major changes to the Town of Ashland's population and housing market in the past five years. Although Ashland resides in Hanover County, the town sees some differences in population and housing trends from the rest of the county. 

```{r}
#| label: setup

library(tidyverse)
library(tidycensus)
library(sf)
library(sp)
library(plotly)
library(simplevis)
library(tidytext)
library(scales)
library(readxl)
library(zoo)
library(lubridate)
library(janitor)
library(viridis)
library(kableExtra)
library(leaflet)
library(FinCal)
library(hdatools)
library(ggiraph)
library(ggtext)


```

```{r}
#| label: fig-ash-map
#| fig-cap: "Map of Town of Ashland"

ashland <- st_read('https://services2.arcgis.com/sKZWgJlU6SekCzQV/arcgis/rest/services/Ashland_Boundary/FeatureServer/0/query?outFields=*&where=1%3D1&f=geojson', quiet = TRUE)

leaflet(ashland) %>% 
  addTiles() %>% 
  addPolygons()

```

## Takeaways

- The Town of Ashland saw recent declines in population, but is still expected to grow --- reaching nearly 10,000 residents by 2050.
- More and more younger adults are coming to the town; mainly renter households with children and homeowners living alone.
- Home prices and rent continue to rise with the influx of higher earning households --- especially higher income renters.
- In contrast to other localities, homeowners are increasingly see cost burden compared to renters who are seeing less.


## Demographic and socioeconomic changes

### Population changes

The gradual population growth that Ashland was seeing throughout the last decade took a change in course between 2019 and 2020. Between these years, it was estimated that Ashland's population declined by 310 individuals. This was a greater decrease than the town's last recorded population decrease between 2010 and 2011, when the town population declined by 38. The small population of Ashland may warrant some caution in describing trends between American Community Survey estimates and decennial census counts.

```{r}
#| label: pop-data
#| eval: false

ashland_pop <- get_estimates(
  geography = "place",
  state = "VA",
  variables = "POP",
  year = 2019,
  time_series = TRUE
) %>% 
  filter(NAME == "Ashland town, Virginia")

ashland_20 <- get_decennial(
  geography = "place",
  state = "VA",
  year = 2020,
  sumfile = "pl",
  variables = "P1_001N"
) %>% 
  filter(NAME == "Ashland town, Virginia")

ashpop_clean <- ashland_pop |>
  filter(!DATE %in% c(2, 3)) |> # Remove non-decennial 2010 counts
  mutate(year = # Translate date codes into years
    case_when(
      DATE == 1 ~ "2010",
      DATE == 4 ~ "2011",
      DATE == 5 ~ "2012",
      DATE == 6 ~ "2013",
      DATE == 7 ~ "2014",
      DATE == 8 ~ "2015",
      DATE == 9 ~ "2016",
      DATE == 10 ~ "2017",
      DATE == 11 ~ "2018",
      DATE == 12 ~ "2019")) |> 
  mutate(counttype = # Add descriptions to count types
      case_when(
        DATE == 1 ~ "Census population",
        TRUE ~ "Population estimate")) |> 
  select( # Simplify columns
    GEOID,
    year,
    counttype,
    value
  )

# Prep total population counts from 2020 Census summary file

census_clean <- ashland_20 |> 
  mutate(year = "2020", # Add year and count type columns
         counttype = "Census population") |>
  select( # Simplify columns
    GEOID,
    year,
    counttype,
    value
  )

population_data <- ashpop_clean |> 
  bind_rows(census_clean)

write_rds(population_data, "data/population_data.rds")

```

```{r}
#| label: fig-ash-pop
#| fig-cap: "Town of Ashland: Total Population"

population_data <- read_rds("data/population_data.rds")

pop_bar <- ggplot(population_data,
                      aes(x = year,
                          y = value,
                          fill = value,
                          data_id = value,
                          tooltip = label_comma()(value))) +
  geom_col() +
  geom_col_interactive(size = 2) +
  scale_fill_gradientn(colours = c("#CAE3C2", "#74BA91", "#519B8D", "#2B4258")) +
  theme_pha() +
  labs(title = "Town of Ashland: Total population",
       subtitle = "2010 to 2020",
       caption = "**Source:** U.S. Census Bureau Decennial Census and American Community Survey.") +
  scale_y_continuous(labels = label_comma()) +
  theme(axis.title = element_blank(),
        legend.position = "none",
        panel.background = element_blank(),
        axis.ticks = element_blank(),
        panel.grid.major.y = element_line(color = "grey95",
                                          size = 0.05)) +
  theme(plot.title.position = "plot",
        plot.caption.position = "plot")  +
  add_zero_line("y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = pop_bar,
         height_svg = 4)
} else {
  
  pop_bar
}

```
Population projections produced by the University of Virginia's Weldon Cooper Center show that the town will continue to grow over the coming decades. By 2050, the Weldon Cooper Center expects Ashland to approach nearly 10,000 residents --- a 27 percent increase over 30 years.

```{r}
#| label: fig-pop-forecasts
#| fig-cap: "Town of Ashland: Population forecast"

years <- c(2020, 2030, 2040, 2050)
estimate <- c(7565, 8142, 8824, 9638)
type <- c("Census", "Forecast", "Forecast", "Forecast")

forecast <- data.frame(year = years,
                       estimate = estimate,
                       type = type)


forecast_plot <- ggplot(forecast,
       aes(x = year,
           y = estimate,
           fill = type,
          data_id = estimate,
          tooltip = label_comma()(estimate))) +
  geom_col() +
  geom_col_interactive(size = 2) +
  scale_fill_pha() +
  theme_pha() +
  scale_y_continuous(labels = label_comma()) +
  labs(title="Town of Ashland: Population projections",
       subtitle="Town population in 2020 and forecast to 2050",
       fill="Estimate type",
       caption="**Source:** University of Virginia Weldon Cooper Center for Public Service")  +
  add_zero_line("y")


if (knitr::is_html_output()) {
  
  girafe(ggobj = forecast_plot,
         height_svg = 4)
} else {
  
  forecast_plot
}
  

```


### Household characteristics

From 2016 to 2020, Ashland has seen most of its household growth among homeowners between the ages of 25 and 44 years old (+342). Declines have mainly been among homeowner householders older than 45 years old (-249), while there has been an increase of renter households among this same age group.

```{r}
#| label: hh-age-data
#| eval: false

years <- 2016:2020

b25007_vars <- load_variables(2020, "acs5") |>
  filter(str_sub(name, end = 6) %in% "B25007")

b25007_raw <- map_dfr(years, function(yr){
  b25007_pull <- get_acs(
    geography = "place",
    state = "VA",
    table = "B25007",
    year = yr,
    survey = "acs5",
    cache_table = TRUE
  ) |>
    mutate(year = yr)
}) %>% 
  filter(NAME == "Ashland town, Virginia")


b25007_vars_cleaned <- b25007_vars |>
  separate(label, into = c("est", "total", "tenure", "age"), sep = "!!") |>
  select(variable = name, tenure, age) |>
  drop_na() |>
  mutate(tenure = case_when(
    tenure == "Owner occupied:" ~ "Homeowner",
    tenure == "Renter occupied:" ~ "Renter"
  ))

b25007_ashland <- b25007_raw |>
  right_join(b25007_vars_cleaned, by = "variable") |>
  select(NAME, GEOID, year, tenure, age, estimate, moe) |>
  mutate(NAME = str_remove_all(NAME, "town, Virginia")) |>
  mutate(age = case_when(
    age == "Householder 15 to 24 years" ~ "Under 25 years old",
    age == "Householder 25 to 34 years" ~ "25 to 44 years old",
    age == "Householder 35 to 44 years" ~ "25 to 44 years old",
    age == "Householder 45 to 54 years" ~ "45 to 64 years old",
    age == "Householder 55 to 59 years" ~ "45 to 64 years old",
    age == "Householder 60 to 64 years" ~ "45 to 64 years old",
    age == "Householder 65 to 74 years" ~ "65 years and over",
    age == "Householder 75 to 84 years" ~ "65 years and over",
    age == "Householder 85 years and over" ~ "65 years and over"
  )) |>
  group_by(year, tenure, age) |>
  summarise(estimate = sum(estimate)) |>
  pivot_wider(names_from = year,
              values_from = estimate) |>
  select(tenure, age, `2016`, `2020`) |>
  transform(change = `2020` - `2016`) |>
  group_by(tenure, age) |>
  summarise(change = sum(change)) |>
  mutate(age = fct_relevel(age, "Under 25 years old")) |> 
  group_by(tenure, age) |> 
  summarise(change = sum(change))

write_rds(b25007_ashland, "data/b25007_ashland.rds")

```

```{r}
#| label: fig-hh-age-plot
#| fig.cap: "Town of Ashland: Change in households by age and tenure"

b25007_ashland <- read_rds("data/b25007_ashland.rds")

age_plot <- ggplot(b25007_ashland,
       aes(y = change, 
           x =tenure, 
           fill = tenure,
           data_id = change,
           tooltip = change)) +
  geom_col(position = "dodge") +
  geom_col_interactive(position = "dodge", size = 2) +
  facet_wrap(~age, nrow = 1) +
  theme_pha() +
  scale_fill_pha() +
  scale_y_continuous(labels = label_comma()) +
  labs(title="Town of Ashland: Change in households by age and tenure",
       subtitle="2016 to 2020",
       fill="Tenure",
       caption = "**Source:** U.S. Census Bureau, American Community Survey, Table B25007.")  +
  add_zero_line("y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = age_plot,
         height_svg = 4)
} else {
  
  age_plot
}
  
```

Households with children have only increased among renters in Hanover (+83) from 2016 to 2020 --- this is contrast to Hanover County as a whole, which is seeing an overall decrease in households with children. But the greatest household type to see growth in the town has been among nonfamily homeowner households (i.e. individuals living alone or with roommates) --- similarly to the county.

```{r}
#| label: hh-child-data
#| eval: false

years <- 2016:2020

b25115_vars <- load_variables(2020, "acs5") |>
  filter(str_sub(name, end = 6) %in% "B25115")

b25115_raw <- map_dfr(years, function(yr){
  b25115_pull <- get_acs(
    geography = "place",
    state = "VA",
    table = "B25115",
    year = yr,
    survey = "acs5"
  ) |>
    mutate(year = yr)
}) %>% 
  filter(NAME == "Ashland town, Virginia")

b25115_vars_cleaned <- b25115_vars |>
  separate(label, into = c("est", "total", "tenure", "family", "type", "rel", "children"), sep = "!!") |>
  select(variable = name, tenure, family, type, rel, children) |>
  mutate(tenure = case_when(
    tenure == "Owner occupied:" ~ "Homeowner",
    tenure == "Renter occupied:" ~ "Renter"
  )) |>
  mutate(family = case_when(
    family == "Family households:" ~ "Family household",
    family == "Nonfamily households" ~ "Nonfamily household"
  )) |>
  mutate(type = case_when(
    type == "Married-couple family:" ~ "Married household",
    rel == "Male householder, no spouse present:" ~ "Single male household",
    rel == "Female householder, no spouse present:" ~ "Single female household",
    family == "Nonfamily household" ~ "Nonfamily household"
  )) |>
  mutate(children = case_when(
    rel == "With own children of the householder under 18 years" ~ "With children",
    rel == "No own children of the householder under 18 years" ~ "No children",
    children == "No own children of the householder under 18 years" ~ "No children",
    children == "With own children of the householder under 18 years" ~ "With children",
    family == "Nonfamily household" ~ "Nonfamily household"
  )) |>
  select(variable, tenure, family, type, children) |>
  drop_na()

b25115_ashland <- b25115_raw |>
  right_join(b25115_vars_cleaned, by = "variable") |>
  select(GEOID, year, tenure, type, children, estimate) |>
  pivot_wider(names_from = year,
              values_from = estimate) |>
  transform(change = `2020` - `2016`) |>
  select(tenure, type, children, change) |>
  group_by(tenure, children) |>
  summarise(change = sum(change)) 

write_rds(b25115_ashland, "data/b25115_ashland.rds")

```

```{r}
#| label: fig-hh-child-plot
#| fig.cap: "Town of Ashland: Change in households with children by tenure"

b25115_ashland <- read_rds("data/b25115_ashland.rds")

ch_plot <- ggplot(b25115_ashland,
       aes(x = tenure, 
           y = change, 
           fill = tenure,
           data_id = change,
           tooltip = label_comma()(change))) +
  geom_col(position = "dodge") + 
  geom_col_interactive(position = "dodge", size = 2) +
  theme_pha() + 
  scale_fill_pha() +
  facet_wrap(~children, nrow = 1, scales = "free_x") +
  scale_y_continuous(labels = label_number(style_positive = "plus", big.mark = ",")) +
  labs(title = "Town of Ashland: Change in households<br>with children by tenure",
       subtitle = "2016 to 2020",
       caption = "**Source:** U.S. Census Bureau, American Community Survey, Table B25115.")  +
  add_zero_line("y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = ch_plot,
         height_svg = 4)
  
} else {
  
  ch_plot
  
  }

```

```{r}
#| label: hh-seniors-data
#| eval: false

years <- 2016:2020

b09020_vars <- load_variables(2020, "acs5") |>
  filter(str_sub(name, end = 6) %in% "B09020")

b09020_raw <- map_dfr(years, function(yr){
  b09020_pull <- get_acs(
    geography = "place",
    state = "VA",
    table = "B09020",
    year = yr,
    survey = "acs5"
  ) |>
    mutate(year = yr)
}) %>% 
  filter(NAME == "Ashland town, Virginia")

b09020_vars_cleaned <- b09020_vars |>
  separate(label, into = c("est", "total", "hhgq", "family", "relationship", "sex", "alone"), sep = "!!") |>
  select(variable = name, hhgq, family, relationship, alone) |>
  filter(!variable %in% c("B09020_001", "B09020_002", "B09020_003", "B09020_005", "B09020_006",
                          "B09020_012", "B09020_013", "B09020_014", "B09020_017")) |>
  mutate(relationship = case_when(
    relationship == "Spouse" ~ "With spouse",
    relationship == "Nonrelatives" ~ "With nonrelatives",
    relationship %in% c("Parent", "Parent-in-law", "Other relatives") ~ "With other relative(s)",
    hhgq == "In group quarters" ~ "Group quarters",
    !is.na(alone) ~ alone,
    TRUE ~ relationship
  )) |>
  mutate(family = case_when(
    family == "In family households:" ~ "Family",
    family == "In nonfamily households:" ~ "Nonfamily",
    hhgq == "In group quarters" ~ "Group quarters"
  )) |>
  mutate(across(.fns = ~str_remove_all(.x, ":"))) |>
  select(1,3,4)

b09020_ashland <- b09020_raw |>
  right_join(b09020_vars_cleaned, by = "variable") |>
  select(GEOID, year, family, relationship, estimate) |>
  group_by(year, family, relationship) |> 
  summarise(estimate = sum(estimate)) |> 
  pivot_wider(names_from = year,
              values_from = estimate) |>
  transform(change = `2020` - `2016`) |>
  select(family, relationship, change) |>
  group_by(family, relationship) |> 
  summarise(change = sum(change)) |> 
  ungroup() |>
  mutate(family = str_replace(family, "Group quarters", " "),
         family = fct_relevel(family, "Family", "Nonfamily", " "))

write_rds(b09020_ashland, "data/b09020_ashland.rds")

```

While the senior population has grown overall in Ashland (+107), that growth has been largely among seniors who live with other relatives (+225) or with a spouse (+52). Seniors who are head of households has declined in the town (-87), as well as the number of seniors living in group quarters (-69). Unlike the rest of the county, Ashland is not seeing major increases in seniors living alone.

```{r}
#| label: fig-hh-seniors-plot
#| fig.cap: "Town of Ashland: Change in senior population by living arrangement"

b09020_ashland <- read_rds("data/b09020_ashland.rds")

snr_plot <- ggplot(b09020_ashland,
       aes(y = reorder_within(relationship, change, family),
           x = change,
           fill = family,
           data_id = change,
           tooltip = change)) +
  geom_col(position = "dodge") +
  geom_col_interactive(position = "dodge", size = 2) +
  theme_pha() +
  scale_fill_pha() +
  scale_y_reordered() +
  scale_x_continuous(labels = label_comma()) +
  facet_grid(rows = vars(family), scales = "free_y", space = "free", switch = "y") +
  #scale_y_discrete(labels = function(x) str_wrap(x, width = 10)) +
  labs(title = "Town of Ashland: Change in senior population<br>by living arrangement",
       subtitle="2016 to 2020",
       fill="Tenure",
       caption = "**Source:** U.S. Census Bureau, American Community Survey, Table B09020.") +
  theme(strip.placement = "outside",
        panel.grid.major.y = element_blank(),
        panel.grid.major.x = element_line(color = "#e2e4e3",
                                          size = 0.05))  +
  add_zero_line("x")


if (knitr::is_html_output()) {
  
  girafe(ggobj = snr_plot,
         height_svg = 4)
  
} else {
  
  snr_plot
  
  }

```

### Income and wages

Household incomes have also shifted in Ashland. Although there still remains an income gap between the average homeowner and renter households, the gap has been narrowing in recent years. In 2016, there was an over \$22,000 difference between homeowner and renter median household incomes. But by 2020, that difference had declined to only \$6,364, as homeowner incomes declined and renter incomes increased. 

```{r}
#| label: hh-inc-data
#| eval: false

years <- 2016:2020

b25119_vars <- load_variables(2020, "acs5") %>% 
 filter(str_sub(name, end = 6) %in% "B25119")

b25119_raw <- map_dfr(years, function(yr){
  b25119_pull <- get_acs(
    geography = "place",
    state = "VA",
    table = "B25119",
    year = yr,
    survey = "acs5",
    cache_table = TRUE
  ) |> 
    mutate(year = yr)
  }) %>% 
  filter(NAME == "Ashland town, Virginia")

b25119_vars_cleaned <- b25119_vars %>% 
  separate(label, into = c("est", "lab", "total", "tenure"), sep = "!!") %>% 
  select(variable = name, tenure) %>% 
  drop_na() %>% 
  mutate(tenure = case_when(
    tenure == "Owner occupied (dollars)" ~ "Homeowner",
    tenure == "Renter occupied (dollars)" ~ "Renter"
  ))

b25119_data <- b25119_raw %>% 
  right_join(b25119_vars_cleaned, by = "variable") %>% 
  select(year, tenure, estimate)

cpi <- read_excel("data/CPI_U_RS.xlsx")
cpi <- cpi |> 
  rename(year = Year,
         priceindex = Index) |> 
  transform(year = as.numeric(year))

b25119_ashland <- b25119_data |> 
  left_join(cpi, by = 'year') |> 
  transform(dollars20 = ((381.2/priceindex)*estimate)) |> 
  select(year, tenure, dollars20, cdollars = estimate) |> 
  select(year, tenure, dollars20, cdollars)

write_rds(b25119_ashland, "data/b25119_ashland.rds")

```


```{r}
#| label: fig-med-inc
#| fig-cap: "Town of Ashland: Median household income by tenure"

b25119_ashland <- read_rds("data/b25119_ashland.rds")

inc_plot <- ggplot(b25119_ashland, 
       aes(x = year,
           y = dollars20,
           fill = tenure,
           data_id = dollars20,
           tooltip = dollar_format()(dollars20))) +
  geom_col() +
  geom_col_interactive(size =2) +
  theme_pha() +
  scale_fill_pha() +
  facet_wrap(~tenure) +
  scale_y_continuous(labels = dollar_format(), limits = c(0, NA)) +
  labs(title = "Town of Ashland: Median houshold income by tenure", 
       color = "Tenure",
       subtitle = "2016 to 2020 | Adjusted to 2020 dollars",
       caption = "**Source:** U.S. Census Bureau, American Community Survey, Table B25119.")  +
  add_zero_line("y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = inc_plot,
         height_svg = 4)
  
} else {
  
  inc_plot
  
  }
```

## Housing supply and market changes

### Homeownership

Median home prices in Ashland had crossed into the \$400,000 range well before the pandemic, but from 2021 onward, median home price has remained above \$400,000 for the majority of the time. Homeownership demand in the town continues to drive prices upwards and the trend is expected to continue.

```{r}
#| label: ash-sales-data

# Data pulled on 2022-10-13

annual_cpi <- read_csv("data/CPIAUCSL_annual.csv") %>% 
  mutate(year = year(DATE)) %>% 
  select(year, cpi = CPIAUCSL)

ashland_medsales <- read_csv("data/ashland_med_sales.csv") %>% 
  clean_names() %>% 
  pivot_longer(cols = starts_with("sale"),
    names_to = "label",
    values_to = "med_price",
    values_drop_na = TRUE
  ) |> 
mutate(year = substr(label, 19,23)) %>% 
  select(year, month, med_price)

ashland_medsales$month <- match(ashland_medsales$month, month.abb)

ashland_medsales <- ashland_medsales %>% 
  mutate(date = make_date(year, month)) %>% 
  mutate(med_price = parse_number(med_price))

cpi <- read_csv("data/CPIAUCSL.csv") %>% 
  select(date = DATE, cpi = CPIAUCSL)

ashland_priceadj <- ashland_medsales %>% 
  left_join(cpi, by = "date") %>% 
  mutate(dollars = ((296.761/cpi)*med_price)) %>%  
  mutate(year = year(date)) %>% 
  select(date, year, med_price, dollars, cpi) %>% 
  filter(year >= 2017)

```

```{r}
#| label: fig-ash-sales
#| fig-cap: "Town of Ashland: Monthly median sales price"

sales_plot <- ggplot(ashland_priceadj,
       aes(x = date,
           y = dollars,
           color = dollars,
           group = 1)) +
  geom_line(size = 1) +
  geom_point_interactive(
    aes(data_id = dollars,
        tooltip = label_dollar()(dollars)),
    size = 0.5) + 
  theme_pha() +
  scale_color_gradientn(colours = c("#CAE3C2", "#74BA91", "#519B8D", "#2B4258")) +
  scale_y_continuous(labels = label_dollar()) +
  scale_x_date(date_labels = "%Y", date_breaks = "1 year") +
  labs(title = "Town of Ashland: Median home sales price",
       subtitle = "January 2017 to September 2022",
       caption = "**Source:** Central Virginia Regional Multiple Listing Service.") +
  geom_smooth(method = "lm", color = "#Be451c")

if (knitr::is_html_output()) {
  
  girafe(ggobj = sales_plot,
         height_svg = 4)
  
} else {
  
  sales_plot
  
  }

```

### Rental

From 2016 Q1 to 2021 Q1, average market asking rent in Ashland has been on a gradual increase --- just a 7 percent increase during this timeframe. But in 2021 onward, the Ashland rental market has seen a dramatic shift upward. From 2021 Q1 to 2022 Q3, rent growth has doubled to 15 percent in a shorter period of time. The median renter household further points to an increasing number of higher income earners demanding rental options in the town.

```{r}
#| label: fig-ash-rent
#| fig-cap: "Town of Ashland: Average asking rent"

rent <- read_csv("data/ashland_rent.csv") %>%  
  pivot_longer(2:3,
               names_to = "rent",
               values_to = "value") %>%  
  mutate(period = as.Date(as.yearqtr(period, format = "%Y Q%q"), frac = 1)) %>%  
  mutate(year = year(period)) %>%  
  filter(year >= 2016) %>%  
  mutate(rent = case_when(
    rent == "rent" ~ "Nominal rent",
    rent == "rent_adj" ~ "Inflation-adjusted rent"
  ))

rent_plot <- ggplot(rent,
                    aes(x = period,
                        y = value, 
                        color = rent)) +
  geom_line(size = 1) +
  geom_point_interactive(
    aes(data_id = value,
        tooltip = label_dollar()(value)),
    size = 0.5) + 
  theme_pha() +
  scale_color_pha() +
  labs(title = "Town of Ashland: Average asking rent",
       subtitle = "2016 Q1 to 2022 Q3",
       color = "Dollar value",
       caption = "**Source:** CoStar Group, Inc.") +
  scale_y_continuous(labels = label_dollar()) +
  scale_x_date(date_labels = "%Y", date_breaks = "2 years", limits = as.Date(c("2016-01-01", "2022-09-30"))) +
  theme(legend.position = "top")

if (knitr::is_html_output()) {
  
  girafe(ggobj = rent_plot,
         height_svg = 4)
  
} else {
  
  rent_plot
  
  }


```

## Gap analysis

### Affordability of current housing stock

In the Town of Ashland, renter median household incomes have been growing --- allowing for the average renter to afford higher rents. In 2020, the typical Ashland renter household could afford a rent of \$1,234 without being cost-burdened. This rent is \$275 more than what the average asking rent in 2020 was. 

Although this suggests that renters in the town are more able to afford rental housing, it could also point to lower income renter households having to move elsewhere in order to afford housing.

```{r}
#| label: fig-ash-gap
#| fig-cap: "Town of Ashland: Rental housing gap"

ashland_rent <- read_csv("data/ashland_annual_rent.csv")

rent_gap <- b25119_ashland %>% 
  filter(tenure == "Renter") %>% 
  left_join(ashland_rent, by = "year") %>% 
  mutate(affordable_rent = (cdollars/12)*.30) %>% 
  select(year, rent, affordable_rent) %>% 
  pivot_longer(2:3, 
               names_to = "label",
               values_to = "rent") %>% 
  mutate(label = case_when(
    label == "rent" ~ "Current rent",
    label == "affordable_rent" ~ "Affordable rent"
  ))

rentgap <- ggplot(rent_gap,
       aes(x = year,
           y = rent,
           fill = label,
           data_id = rent, 
           tooltip = dollar_format()(rent))) +
  geom_col(position = "dodge") +
  geom_col_interactive(position = "dodge", size =2) +
  theme_pha() +
  scale_fill_pha() +
  scale_y_continuous(labels = label_dollar()) +
  labs(title = "Town of Ashland: Average asking rent versus<br>rent affordable to median renter income",
       subtitle = "2016 to 2020",
       fill = "Nominal rent",
       caption = "**Sources:** U.S. Census Bureau, American Community Survey, Table B25119<br>and CoStar Group Inc.") +
  theme(legend.position = "top")  +
  add_zero_line("y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = rentgap,
         height_svg = 4)
  
} else {
  
  rentgap
  
  }


```

In spite of increasing incomes among renters in the town, the high price of homeownership continues to keep renters from moving to more permanent tenureship. In 2016. the gap between income needed to afford the median home price and the typical renter income was just over \$17,000. By 2020, that gap had only decreased by about \$2,000 --- leaving the typical renter \$15,241 away from being able to afford the median home price in 2020 (\$305,000).

```{r}
#| label: fig-sales-gap
#| fig-cap: "Town of Ashland: Income needed to afford median home price versus median renter income"


ashland_medsales <- read_csv("data/ashland_mls_medsales.csv")

int_annual <- read_csv("data/fredmac_int_annual.csv")

downpayment <- 0.05 # 5% downpayment
closingcosts <- 0.015 # 1.5% closing costs
utilities <- 250 # Assume $250/month for utilities

slgap <- b25119_ashland %>% 
  filter(tenure == "Renter") %>% 
  left_join(ashland_medsales, by = "year") %>% 
  left_join(int_annual, by = "year") |> 
  drop_na() |> 
  mutate(principal = price - (price*downpayment)) |> 
  mutate(loanamt = principal/(1-closingcosts)) |> 
  mutate(mortgage = abs(pmt((annual_int/12), 360, loanamt, 0)) + 250) |> 
  mutate(inc_needed = ((mortgage*10)/2.8)*12) |> 
  select(year, inc_needed, cdollars) %>% 
  pivot_longer(2:3,
               names_to = "label",
               values_to = "dollars") |> 
  mutate(label = case_when(
    label == "cdollars" ~ "Renter median household income",
    label == "inc_needed" ~ "Income needed to afford median home price"
  ))
  
slgap_plot <- ggplot(slgap,
       aes(x = year,
           y = dollars,
           fill = str_wrap(label, 25),
           data_id = dollars,
           tooltip = dollar_format()(dollars))) +
  geom_col(position = "dodge") +
  geom_col_interactive(position = "dodge", size = 2) +
  scale_y_continuous(labels = label_dollar()) + 
  theme_pha() +
  scale_fill_pha() +
labs(title = "Town of Ashland: Income needed to afford median home price<br>versus median renter income",
       subtitle = "2016 to 2020",
       fill = "Nominal income",
       caption = "**Sources:** U.S. Census Bureau, American Community Survey, Table B25119<br>and Central Virginia Regional Multiple Listing Service.") +
  theme(legend.position = "top")  +
  add_zero_line("y")

if (knitr::is_html_output()) {
  
  girafe(ggobj = slgap_plot,
         height_svg = 4) 
  
} else {
  
  slgap_plot
  
  }

```
In contrast to other localities, homeowner cost burden has increased in Ashland. From 2015 to 2019, there has been an increase of 57 cost-burdened homeowner households --- a 19 percent increase. However, the number of cost-burdened renter households decreased from 690 to 580 --- a 16 percent decrease.
```{r}
#| label: fig-ash-cb
#| fig-cap: "Town of Ashland: Cost burdened households by tenure"

ashland_cb <- read_csv("data/ashland_cb.csv")

ashland_cb_cln <- ashland_cb %>% 
  group_by(year, grouping) %>% 
  summarise(Homeowner = sum(Homeowner),
            Renter = sum(Renter)) %>% 
  pivot_longer(3:4,
               names_to = "tenure",
               values_to = "estimate") %>% 
  mutate(year = as.character(year)) %>% 
  filter(grouping == "Cost-burdened")

cb_plot <- ggplot(ashland_cb_cln,
       aes(x = year,
           y = estimate,
           fill = tenure,
           data_id = estimate,
           tooltip = estimate)) +
  geom_col() +
  geom_col_interactive(size = 2) +
  theme_pha() + 
  scale_fill_pha() +
  facet_wrap(~tenure) +
  labs(title = "Town of Ashland: Cost burdened households by tenure",
       subtitle = "2015 and 2019",
       fill = "Cost burden") +
  scale_y_continuous(labels = number_format(big.mark = ",")) +
  theme(axis.title = element_blank(),
        legend.position = "none")  +
  add_zero_line("y")
  
if (knitr::is_html_output()) {
  
  girafe(ggobj = cb_plot,
         height_svg = 4) 
  
} else {
  
  cb_plot
  
  }
  

```
